fix: circular dependency, util.ts import intersect.ts, intersect.ts import util.ts
