+++ docs/assets/search.js
@@ -1,1 +1,1 @@
-window.searchData = JSON.parse("{\"kinds\":{\"8\":\"Enumeration\",\"16\":\"Enumeration Member\",\"32\":\"Variable\",\"64\":\"Function\",\"128\":\"Class\",\"256\":\"Interface\",\"512\":\"Constructor\",\"1024\":\"Property\",\"2048\":\"Method\",\"65536\":\"Type literal\",\"262144\":\"Accessor\",\"4194304\":\"Type alias\"},\"rows\":[{\"kind\":4194304,\"name\":\"DecompPolygon\",\"url\":\"types/DecompPolygon.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"DecompPoint\",\"url\":\"types/DecompPoint.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":8,\"name\":\"BodyType\",\"url\":\"enums/BodyType.html\",\"classes\":\"tsd-kind-enum\"},{\"kind\":16,\"name\":\"Ellipse\",\"url\":\"enums/BodyType.html#Ellipse\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Line\",\"url\":\"enums/BodyType.html#Line\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Circle\",\"url\":\"enums/BodyType.html#Circle\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Box\",\"url\":\"enums/BodyType.html#Box\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Point\",\"url\":\"enums/BodyType.html#Point\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Polygon\",\"url\":\"enums/BodyType.html#Polygon\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":4194304,\"name\":\"Leaf\",\"url\":\"types/Leaf.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":256,\"name\":\"ChildrenData\",\"url\":\"interfaces/ChildrenData.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"children\",\"url\":\"interfaces/ChildrenData.html#children\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ChildrenData\"},{\"kind\":256,\"name\":\"Data\",\"url\":\"interfaces/Data.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"data\",\"url\":\"interfaces/Data.html#data\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Data\"},{\"kind\":256,\"name\":\"BodyOptions\",\"url\":\"interfaces/BodyOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"interfaces/BodyOptions.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"interfaces/BodyOptions.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":1024,\"name\":\"isCentered\",\"url\":\"interfaces/BodyOptions.html#isCentered\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"interfaces/BodyOptions.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"interfaces/BodyOptions.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":256,\"name\":\"RaycastHit\",\"url\":\"interfaces/RaycastHit.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"point\",\"url\":\"interfaces/RaycastHit.html#point\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"RaycastHit\"},{\"kind\":1024,\"name\":\"body\",\"url\":\"interfaces/RaycastHit.html#body\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"RaycastHit\"},{\"kind\":256,\"name\":\"PotentialVector\",\"url\":\"interfaces/PotentialVector.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"x\",\"url\":\"interfaces/PotentialVector.html#x\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"PotentialVector\"},{\"kind\":1024,\"name\":\"y\",\"url\":\"interfaces/PotentialVector.html#y\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"PotentialVector\"},{\"kind\":256,\"name\":\"Vector\",\"url\":\"interfaces/Vector.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"x\",\"url\":\"interfaces/Vector.html#x\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Vector\"},{\"kind\":1024,\"name\":\"y\",\"url\":\"interfaces/Vector.html#y\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Vector\"},{\"kind\":256,\"name\":\"GetAABBAsBox\",\"url\":\"interfaces/GetAABBAsBox.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":2048,\"name\":\"getAABBAsBox\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"GetAABBAsBox\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox.getAABBAsBox-1.__type\",\"classes\":\"tsd-kind-type-literal\",\"parent\":\"GetAABBAsBox.getAABBAsBox.getAABBAsBox\"},{\"kind\":1024,\"name\":\"pos\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox.getAABBAsBox-1.__type.pos\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"GetAABBAsBox.getAABBAsBox.getAABBAsBox.__type\"},{\"kind\":1024,\"name\":\"w\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox.getAABBAsBox-1.__type.w\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"GetAABBAsBox.getAABBAsBox.getAABBAsBox.__type\"},{\"kind\":1024,\"name\":\"h\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox.getAABBAsBox-1.__type.h\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"GetAABBAsBox.getAABBAsBox.getAABBAsBox.__type\"},{\"kind\":4194304,\"name\":\"Body\",\"url\":\"types/Body.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":256,\"name\":\"BodyProps\",\"url\":\"interfaces/BodyProps.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"interfaces/BodyProps.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"interfaces/BodyProps.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"interfaces/BodyProps.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"offset\",\"url\":\"interfaces/BodyProps.html#offset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"interfaces/BodyProps.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"interfaces/BodyProps.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"interfaces/BodyProps.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"interfaces/BodyProps.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"interfaces/BodyProps.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"setAngle\",\"url\":\"interfaces/BodyProps.html#setAngle\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"setOffset\",\"url\":\"interfaces/BodyProps.html#setOffset\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"interfaces/BodyProps.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"interfaces/BodyProps.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"interfaces/BodyProps.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"interfaces/BodyProps.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"isCentered\",\"url\":\"interfaces/BodyProps.html#isCentered\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"interfaces/BodyProps.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"interfaces/BodyProps.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":4194304,\"name\":\"SATTest\",\"url\":\"types/SATTest.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/SATTest.html#__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-type-alias\",\"parent\":\"SATTest\"},{\"kind\":128,\"name\":\"Circle\",\"url\":\"classes/Circle.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Circle.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Circle.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Circle.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Circle.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Circle.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Circle.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"offset\",\"url\":\"classes/Circle.html#offset\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"offsetCopy\",\"url\":\"classes/Circle.html#offsetCopy\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Circle.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"classes/Circle.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Circle.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Circle.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Circle.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Circle.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Circle.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"isCentered\",\"url\":\"classes/Circle.html#isCentered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"unscaledRadius\",\"url\":\"classes/Circle.html#unscaledRadius\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Circle.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Circle.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Circle.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Circle.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Circle.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Circle.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Circle.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"setAngle\",\"url\":\"classes/Circle.html#setAngle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"setOffset\",\"url\":\"classes/Circle.html#setOffset\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Circle.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Circle.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"getOffsetWithAngle\",\"url\":\"classes/Circle.html#getOffsetWithAngle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Circle\"},{\"kind\":128,\"name\":\"Ellipse\",\"url\":\"classes/Ellipse.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Ellipse.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Ellipse.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Ellipse.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"_radiusX\",\"url\":\"classes/Ellipse.html#_radiusX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"_radiusY\",\"url\":\"classes/Ellipse.html#_radiusY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"_step\",\"url\":\"classes/Ellipse.html#_step\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Ellipse.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"step\",\"url\":\"classes/Ellipse.html#step\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"radiusX\",\"url\":\"classes/Ellipse.html#radiusX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"radiusY\",\"url\":\"classes/Ellipse.html#radiusY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"center\",\"url\":\"classes/Ellipse.html#center\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Ellipse.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Ellipse.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Ellipse.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Ellipse.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Ellipse.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Ellipse.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Ellipse.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Ellipse.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Ellipse.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Ellipse.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Ellipse.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Ellipse.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Ellipse.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Ellipse.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Ellipse.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Ellipse.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Ellipse.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Ellipse.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Ellipse.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Ellipse.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Ellipse.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Ellipse.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Ellipse.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Ellipse.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Ellipse.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Ellipse.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Ellipse.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Ellipse.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Ellipse.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Ellipse.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Ellipse.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Ellipse.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":64,\"name\":\"isSimple\",\"url\":\"functions/isSimple.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":128,\"name\":\"Polygon\",\"url\":\"classes/Polygon.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Polygon.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Polygon.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Polygon.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Polygon.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Polygon.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Polygon.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Polygon.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Polygon.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Polygon.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Polygon.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Polygon.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Polygon.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Polygon.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Polygon.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Polygon.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Polygon.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Polygon.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Polygon.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Polygon.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Polygon.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Polygon.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Polygon.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Polygon.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Polygon.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Polygon.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Polygon.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Polygon.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Polygon.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Polygon.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Polygon.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Polygon.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Polygon.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Polygon.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Polygon.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Polygon.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Polygon.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":128,\"name\":\"Box\",\"url\":\"classes/Box.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Box.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Box.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Box.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"_width\",\"url\":\"classes/Box.html#_width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"_height\",\"url\":\"classes/Box.html#_height\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"width\",\"url\":\"classes/Box.html#width\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"height\",\"url\":\"classes/Box.html#height\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Box.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Box.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Box.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Box.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Box.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Box.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Box.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Box.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Box.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Box.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Box.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Box.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Box.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Box.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Box.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Box.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Box.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Box.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Box.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Box.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Box.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Box.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Box.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Box.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Box.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Box.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Box.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Box.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Box.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Box.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Box.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Box.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Box.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":128,\"name\":\"Point\",\"url\":\"classes/Point.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Point.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Point.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Point.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"_width\",\"url\":\"classes/Point.html#_width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"_height\",\"url\":\"classes/Point.html#_height\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"width\",\"url\":\"classes/Point.html#width\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"height\",\"url\":\"classes/Point.html#height\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Point.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Point.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Point.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Point.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Point.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Point.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Point.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Point.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Point.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Point.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Point.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Point.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Point.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Point.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Point.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Point.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Point.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Point.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Point.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Point.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Point.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Point.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Point.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Point.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Point.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Point.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Point.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Point.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Point.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Point.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Point.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Point.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Point.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":128,\"name\":\"Line\",\"url\":\"classes/Line.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Line.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Line.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Line.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"start\",\"url\":\"classes/Line.html#start\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"end\",\"url\":\"classes/Line.html#end\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"getCentroid\",\"url\":\"classes/Line.html#getCentroid\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Line.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Line.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Line.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Line.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Line.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Line.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Line.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Line.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Line.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Line.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Line.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Line.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Line.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Line.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Line.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Line.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Line.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Line.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Line.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Line.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Line.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Line.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Line.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Line.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Line.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Line.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Line.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Line.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Line.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Line.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Line.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Line.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Line.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":128,\"name\":\"System\",\"url\":\"classes/System.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":1024,\"name\":\"response\",\"url\":\"classes/System.html#response\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":1024,\"name\":\"ray\",\"url\":\"classes/System.html#ray\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"remove\",\"url\":\"classes/System.html#remove\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"insert\",\"url\":\"classes/System.html#insert\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/System.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"update\",\"url\":\"classes/System.html#update\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"separate\",\"url\":\"classes/System.html#separate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"checkOne\",\"url\":\"classes/System.html#checkOne\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"checkAll\",\"url\":\"classes/System.html#checkAll\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"getPotentials\",\"url\":\"classes/System.html#getPotentials\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"checkCollision\",\"url\":\"classes/System.html#checkCollision\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"raycast\",\"url\":\"classes/System.html#raycast\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"traverse\",\"url\":\"classes/System.html#traverse\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":1024,\"name\":\"data\",\"url\":\"classes/System.html#data\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/System.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"drawBVH\",\"url\":\"classes/System.html#drawBVH\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createPoint\",\"url\":\"classes/System.html#createPoint\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createLine\",\"url\":\"classes/System.html#createLine\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createCircle\",\"url\":\"classes/System.html#createCircle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createBox\",\"url\":\"classes/System.html#createBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createEllipse\",\"url\":\"classes/System.html#createEllipse\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createPolygon\",\"url\":\"classes/System.html#createPolygon\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":64,\"name\":\"deg2rad\",\"url\":\"functions/deg2rad-1.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"rad2deg\",\"url\":\"functions/rad2deg-1.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"createEllipse\",\"url\":\"functions/createEllipse.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"createBox\",\"url\":\"functions/createBox.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"ensureVectorPoint\",\"url\":\"functions/ensureVectorPoint.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"ensurePolygonPoints\",\"url\":\"functions/ensurePolygonPoints.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"distance\",\"url\":\"functions/distance.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"clockwise\",\"url\":\"functions/clockwise.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"extendBody\",\"url\":\"functions/extendBody.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"bodyMoved\",\"url\":\"functions/bodyMoved.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"notIntersectAABB\",\"url\":\"functions/notIntersectAABB.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectAABB\",\"url\":\"functions/intersectAABB.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"checkAInB\",\"url\":\"functions/checkAInB.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"clonePointsArray\",\"url\":\"functions/clonePointsArray.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"dashLineTo\",\"url\":\"functions/dashLineTo.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"mapVectorToArray\",\"url\":\"functions/mapVectorToArray.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"mapArrayToVector\",\"url\":\"functions/mapArrayToVector.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"ensureConvex\",\"url\":\"functions/ensureConvex.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"getBounceDirection\",\"url\":\"functions/getBounceDirection.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"getSATTest\",\"url\":\"functions/getSATTest.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"drawPolygon\",\"url\":\"functions/drawPolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":32,\"name\":\"DEG2RAD\",\"url\":\"variables/DEG2RAD.html\",\"classes\":\"tsd-kind-variable\"},{\"kind\":32,\"name\":\"RAD2DEG\",\"url\":\"variables/RAD2DEG.html\",\"classes\":\"tsd-kind-variable\"},{\"kind\":64,\"name\":\"polygonInCircle\",\"url\":\"functions/polygonInCircle.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"pointInPolygon\",\"url\":\"functions/pointInPolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"polygonInPolygon\",\"url\":\"functions/polygonInPolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"pointOnCircle\",\"url\":\"functions/pointOnCircle.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"circleInCircle\",\"url\":\"functions/circleInCircle.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"circleInPolygon\",\"url\":\"functions/circleInPolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"circleOutsidePolygon\",\"url\":\"functions/circleOutsidePolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectLineCircle\",\"url\":\"functions/intersectLineCircle.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectLineLineFast\",\"url\":\"functions/intersectLineLineFast.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectLineLine\",\"url\":\"functions/intersectLineLine.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectLinePolygon\",\"url\":\"functions/intersectLinePolygon.html\",\"classes\":\"tsd-kind-function\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,54.496]],[\"comment/0\",[]],[\"name/1\",[1,54.496]],[\"comment/1\",[]],[\"name/2\",[2,54.496]],[\"comment/2\",[]],[\"name/3\",[3,49.388]],[\"comment/3\",[]],[\"name/4\",[4,49.388]],[\"comment/4\",[]],[\"name/5\",[5,49.388]],[\"comment/5\",[]],[\"name/6\",[6,49.388]],[\"comment/6\",[]],[\"name/7\",[7,46.023]],[\"comment/7\",[]],[\"name/8\",[8,49.388]],[\"comment/8\",[]],[\"name/9\",[9,54.496]],[\"comment/9\",[]],[\"name/10\",[10,54.496]],[\"comment/10\",[]],[\"name/11\",[11,54.496]],[\"comment/11\",[]],[\"name/12\",[12,46.023]],[\"comment/12\",[]],[\"name/13\",[12,46.023]],[\"comment/13\",[]],[\"name/14\",[13,54.496]],[\"comment/14\",[]],[\"name/15\",[14,37.15]],[\"comment/15\",[]],[\"name/16\",[15,37.15]],[\"comment/16\",[]],[\"name/17\",[16,37.15]],[\"comment/17\",[]],[\"name/18\",[17,46.023]],[\"comment/18\",[]],[\"name/19\",[18,37.15]],[\"comment/19\",[]],[\"name/20\",[19,54.496]],[\"comment/20\",[]],[\"name/21\",[7,46.023]],[\"comment/21\",[]],[\"name/22\",[20,49.388]],[\"comment/22\",[]],[\"name/23\",[21,54.496]],[\"comment/23\",[]],[\"name/24\",[22,37.15]],[\"comment/24\",[]],[\"name/25\",[23,37.15]],[\"comment/25\",[]],[\"name/26\",[24,54.496]],[\"comment/26\",[]],[\"name/27\",[22,37.15]],[\"comment/27\",[]],[\"name/28\",[23,37.15]],[\"comment/28\",[]],[\"name/29\",[25,49.388]],[\"comment/29\",[]],[\"name/30\",[25,49.388]],[\"comment/30\",[]],[\"name/31\",[26,49.388]],[\"comment/31\",[]],[\"name/32\",[27,54.496]],[\"comment/32\",[]],[\"name/33\",[28,54.496]],[\"comment/33\",[]],[\"name/34\",[29,54.496]],[\"comment/34\",[]],[\"name/35\",[20,49.388]],[\"comment/35\",[]],[\"name/36\",[30,54.496]],[\"comment/36\",[]],[\"name/37\",[31,38.402]],[\"comment/37\",[]],[\"name/38\",[32,38.402]],[\"comment/38\",[]],[\"name/39\",[33,38.402]],[\"comment/39\",[]],[\"name/40\",[34,49.388]],[\"comment/40\",[]],[\"name/41\",[35,37.15]],[\"comment/41\",[]],[\"name/42\",[36,38.402]],[\"comment/42\",[]],[\"name/43\",[37,38.402]],[\"comment/43\",[]],[\"name/44\",[38,38.402]],[\"comment/44\",[]],[\"name/45\",[39,38.402]],[\"comment/45\",[]],[\"name/46\",[40,49.388]],[\"comment/46\",[]],[\"name/47\",[41,49.388]],[\"comment/47\",[]],[\"name/48\",[42,37.15]],[\"comment/48\",[]],[\"name/49\",[43,38.402]],[\"comment/49\",[]],[\"name/50\",[14,37.15]],[\"comment/50\",[]],[\"name/51\",[15,37.15]],[\"comment/51\",[]],[\"name/52\",[16,37.15]],[\"comment/52\",[]],[\"name/53\",[17,46.023]],[\"comment/53\",[]],[\"name/54\",[18,37.15]],[\"comment/54\",[]],[\"name/55\",[44,54.496]],[\"comment/55\",[]],[\"name/56\",[26,49.388]],[\"comment/56\",[]],[\"name/57\",[5,49.388]],[\"comment/57\",[]],[\"name/58\",[45,39.833]],[\"comment/58\",[]],[\"name/59\",[46,39.833]],[\"comment/59\",[]],[\"name/60\",[47,39.833]],[\"comment/60\",[]],[\"name/61\",[48,39.833]],[\"comment/61\",[]],[\"name/62\",[49,39.833]],[\"comment/62\",[]],[\"name/63\",[33,38.402]],[\"comment/63\",[]],[\"name/64\",[34,49.388]],[\"comment/64\",[]],[\"name/65\",[50,54.496]],[\"comment/65\",[]],[\"name/66\",[18,37.15]],[\"comment/66\",[]],[\"name/67\",[17,46.023]],[\"comment/67\",[]],[\"name/68\",[14,37.15]],[\"comment/68\",[]],[\"name/69\",[15,37.15]],[\"comment/69\",[]],[\"name/70\",[35,37.15]],[\"comment/70\",[]],[\"name/71\",[32,38.402]],[\"comment/71\",[]],[\"name/72\",[31,38.402]],[\"comment/72\",[]],[\"name/73\",[16,37.15]],[\"comment/73\",[]],[\"name/74\",[51,54.496]],[\"comment/74\",[]],[\"name/75\",[22,37.15]],[\"comment/75\",[]],[\"name/76\",[23,37.15]],[\"comment/76\",[]],[\"name/77\",[52,39.833]],[\"comment/77\",[]],[\"name/78\",[36,38.402]],[\"comment/78\",[]],[\"name/79\",[37,38.402]],[\"comment/79\",[]],[\"name/80\",[38,38.402]],[\"comment/80\",[]],[\"name/81\",[39,38.402]],[\"comment/81\",[]],[\"name/82\",[40,49.388]],[\"comment/82\",[]],[\"name/83\",[41,49.388]],[\"comment/83\",[]],[\"name/84\",[43,38.402]],[\"comment/84\",[]],[\"name/85\",[42,37.15]],[\"comment/85\",[]],[\"name/86\",[53,54.496]],[\"comment/86\",[]],[\"name/87\",[3,49.388]],[\"comment/87\",[]],[\"name/88\",[45,39.833]],[\"comment/88\",[]],[\"name/89\",[31,38.402]],[\"comment/89\",[]],[\"name/90\",[32,38.402]],[\"comment/90\",[]],[\"name/91\",[54,54.496]],[\"comment/91\",[]],[\"name/92\",[55,54.496]],[\"comment/92\",[]],[\"name/93\",[56,54.496]],[\"comment/93\",[]],[\"name/94\",[16,37.15]],[\"comment/94\",[]],[\"name/95\",[57,54.496]],[\"comment/95\",[]],[\"name/96\",[58,54.496]],[\"comment/96\",[]],[\"name/97\",[59,54.496]],[\"comment/97\",[]],[\"name/98\",[60,54.496]],[\"comment/98\",[]],[\"name/99\",[61,41.503]],[\"comment/99\",[]],[\"name/100\",[46,39.833]],[\"comment/100\",[]],[\"name/101\",[47,39.833]],[\"comment/101\",[]],[\"name/102\",[48,39.833]],[\"comment/102\",[]],[\"name/103\",[49,39.833]],[\"comment/103\",[]],[\"name/104\",[33,38.402]],[\"comment/104\",[]],[\"name/105\",[62,41.503]],[\"comment/105\",[]],[\"name/106\",[18,37.15]],[\"comment/106\",[]],[\"name/107\",[14,37.15]],[\"comment/107\",[]],[\"name/108\",[15,37.15]],[\"comment/108\",[]],[\"name/109\",[35,37.15]],[\"comment/109\",[]],[\"name/110\",[63,41.503]],[\"comment/110\",[]],[\"name/111\",[64,41.503]],[\"comment/111\",[]],[\"name/112\",[65,41.503]],[\"comment/112\",[]],[\"name/113\",[22,37.15]],[\"comment/113\",[]],[\"name/114\",[23,37.15]],[\"comment/114\",[]],[\"name/115\",[36,38.402]],[\"comment/115\",[]],[\"name/116\",[37,38.402]],[\"comment/116\",[]],[\"name/117\",[52,39.833]],[\"comment/117\",[]],[\"name/118\",[38,38.402]],[\"comment/118\",[]],[\"name/119\",[39,38.402]],[\"comment/119\",[]],[\"name/120\",[43,38.402]],[\"comment/120\",[]],[\"name/121\",[42,37.15]],[\"comment/121\",[]],[\"name/122\",[66,41.503]],[\"comment/122\",[]],[\"name/123\",[67,41.503]],[\"comment/123\",[]],[\"name/124\",[68,41.503]],[\"comment/124\",[]],[\"name/125\",[69,41.503]],[\"comment/125\",[]],[\"name/126\",[70,39.833]],[\"comment/126\",[]],[\"name/127\",[71,41.503]],[\"comment/127\",[]],[\"name/128\",[72,41.503]],[\"comment/128\",[]],[\"name/129\",[73,41.503]],[\"comment/129\",[]],[\"name/130\",[74,39.833]],[\"comment/130\",[]],[\"name/131\",[70,39.833]],[\"comment/131\",[]],[\"name/132\",[8,49.388]],[\"comment/132\",[]],[\"name/133\",[45,39.833]],[\"comment/133\",[]],[\"name/134\",[46,39.833]],[\"comment/134\",[]],[\"name/135\",[47,39.833]],[\"comment/135\",[]],[\"name/136\",[48,39.833]],[\"comment/136\",[]],[\"name/137\",[49,39.833]],[\"comment/137\",[]],[\"name/138\",[33,38.402]],[\"comment/138\",[]],[\"name/139\",[32,38.402]],[\"comment/139\",[]],[\"name/140\",[62,41.503]],[\"comment/140\",[]],[\"name/141\",[18,37.15]],[\"comment/141\",[]],[\"name/142\",[14,37.15]],[\"comment/142\",[]],[\"name/143\",[15,37.15]],[\"comment/143\",[]],[\"name/144\",[35,37.15]],[\"comment/144\",[]],[\"name/145\",[31,38.402]],[\"comment/145\",[]],[\"name/146\",[63,41.503]],[\"comment/146\",[]],[\"name/147\",[64,41.503]],[\"comment/147\",[]],[\"name/148\",[65,41.503]],[\"comment/148\",[]],[\"name/149\",[16,37.15]],[\"comment/149\",[]],[\"name/150\",[22,37.15]],[\"comment/150\",[]],[\"name/151\",[23,37.15]],[\"comment/151\",[]],[\"name/152\",[36,38.402]],[\"comment/152\",[]],[\"name/153\",[37,38.402]],[\"comment/153\",[]],[\"name/154\",[52,39.833]],[\"comment/154\",[]],[\"name/155\",[38,38.402]],[\"comment/155\",[]],[\"name/156\",[39,38.402]],[\"comment/156\",[]],[\"name/157\",[43,38.402]],[\"comment/157\",[]],[\"name/158\",[42,37.15]],[\"comment/158\",[]],[\"name/159\",[66,41.503]],[\"comment/159\",[]],[\"name/160\",[67,41.503]],[\"comment/160\",[]],[\"name/161\",[68,41.503]],[\"comment/161\",[]],[\"name/162\",[69,41.503]],[\"comment/162\",[]],[\"name/163\",[70,39.833]],[\"comment/163\",[]],[\"name/164\",[71,41.503]],[\"comment/164\",[]],[\"name/165\",[72,41.503]],[\"comment/165\",[]],[\"name/166\",[73,41.503]],[\"comment/166\",[]],[\"name/167\",[61,41.503]],[\"comment/167\",[]],[\"name/168\",[74,39.833]],[\"comment/168\",[]],[\"name/169\",[6,49.388]],[\"comment/169\",[]],[\"name/170\",[45,39.833]],[\"comment/170\",[]],[\"name/171\",[31,38.402]],[\"comment/171\",[]],[\"name/172\",[32,38.402]],[\"comment/172\",[]],[\"name/173\",[75,49.388]],[\"comment/173\",[]],[\"name/174\",[76,49.388]],[\"comment/174\",[]],[\"name/175\",[77,49.388]],[\"comment/175\",[]],[\"name/176\",[78,49.388]],[\"comment/176\",[]],[\"name/177\",[61,41.503]],[\"comment/177\",[]],[\"name/178\",[46,39.833]],[\"comment/178\",[]],[\"name/179\",[47,39.833]],[\"comment/179\",[]],[\"name/180\",[48,39.833]],[\"comment/180\",[]],[\"name/181\",[49,39.833]],[\"comment/181\",[]],[\"name/182\",[33,38.402]],[\"comment/182\",[]],[\"name/183\",[62,41.503]],[\"comment/183\",[]],[\"name/184\",[18,37.15]],[\"comment/184\",[]],[\"name/185\",[14,37.15]],[\"comment/185\",[]],[\"name/186\",[15,37.15]],[\"comment/186\",[]],[\"name/187\",[35,37.15]],[\"comment/187\",[]],[\"name/188\",[63,41.503]],[\"comment/188\",[]],[\"name/189\",[64,41.503]],[\"comment/189\",[]],[\"name/190\",[65,41.503]],[\"comment/190\",[]],[\"name/191\",[16,37.15]],[\"comment/191\",[]],[\"name/192\",[22,37.15]],[\"comment/192\",[]],[\"name/193\",[23,37.15]],[\"comment/193\",[]],[\"name/194\",[36,38.402]],[\"comment/194\",[]],[\"name/195\",[37,38.402]],[\"comment/195\",[]],[\"name/196\",[52,39.833]],[\"comment/196\",[]],[\"name/197\",[38,38.402]],[\"comment/197\",[]],[\"name/198\",[39,38.402]],[\"comment/198\",[]],[\"name/199\",[43,38.402]],[\"comment/199\",[]],[\"name/200\",[42,37.15]],[\"comment/200\",[]],[\"name/201\",[66,41.503]],[\"comment/201\",[]],[\"name/202\",[67,41.503]],[\"comment/202\",[]],[\"name/203\",[68,41.503]],[\"comment/203\",[]],[\"name/204\",[69,41.503]],[\"comment/204\",[]],[\"name/205\",[70,39.833]],[\"comment/205\",[]],[\"name/206\",[71,41.503]],[\"comment/206\",[]],[\"name/207\",[72,41.503]],[\"comment/207\",[]],[\"name/208\",[73,41.503]],[\"comment/208\",[]],[\"name/209\",[74,39.833]],[\"comment/209\",[]],[\"name/210\",[7,46.023]],[\"comment/210\",[]],[\"name/211\",[45,39.833]],[\"comment/211\",[]],[\"name/212\",[31,38.402]],[\"comment/212\",[]],[\"name/213\",[32,38.402]],[\"comment/213\",[]],[\"name/214\",[75,49.388]],[\"comment/214\",[]],[\"name/215\",[76,49.388]],[\"comment/215\",[]],[\"name/216\",[77,49.388]],[\"comment/216\",[]],[\"name/217\",[78,49.388]],[\"comment/217\",[]],[\"name/218\",[61,41.503]],[\"comment/218\",[]],[\"name/219\",[46,39.833]],[\"comment/219\",[]],[\"name/220\",[47,39.833]],[\"comment/220\",[]],[\"name/221\",[48,39.833]],[\"comment/221\",[]],[\"name/222\",[49,39.833]],[\"comment/222\",[]],[\"name/223\",[33,38.402]],[\"comment/223\",[]],[\"name/224\",[62,41.503]],[\"comment/224\",[]],[\"name/225\",[18,37.15]],[\"comment/225\",[]],[\"name/226\",[14,37.15]],[\"comment/226\",[]],[\"name/227\",[15,37.15]],[\"comment/227\",[]],[\"name/228\",[35,37.15]],[\"comment/228\",[]],[\"name/229\",[63,41.503]],[\"comment/229\",[]],[\"name/230\",[64,41.503]],[\"comment/230\",[]],[\"name/231\",[65,41.503]],[\"comment/231\",[]],[\"name/232\",[16,37.15]],[\"comment/232\",[]],[\"name/233\",[22,37.15]],[\"comment/233\",[]],[\"name/234\",[23,37.15]],[\"comment/234\",[]],[\"name/235\",[36,38.402]],[\"comment/235\",[]],[\"name/236\",[37,38.402]],[\"comment/236\",[]],[\"name/237\",[52,39.833]],[\"comment/237\",[]],[\"name/238\",[38,38.402]],[\"comment/238\",[]],[\"name/239\",[39,38.402]],[\"comment/239\",[]],[\"name/240\",[43,38.402]],[\"comment/240\",[]],[\"name/241\",[42,37.15]],[\"comment/241\",[]],[\"name/242\",[66,41.503]],[\"comment/242\",[]],[\"name/243\",[67,41.503]],[\"comment/243\",[]],[\"name/244\",[68,41.503]],[\"comment/244\",[]],[\"name/245\",[69,41.503]],[\"comment/245\",[]],[\"name/246\",[70,39.833]],[\"comment/246\",[]],[\"name/247\",[71,41.503]],[\"comment/247\",[]],[\"name/248\",[72,41.503]],[\"comment/248\",[]],[\"name/249\",[73,41.503]],[\"comment/249\",[]],[\"name/250\",[74,39.833]],[\"comment/250\",[]],[\"name/251\",[4,49.388]],[\"comment/251\",[]],[\"name/252\",[45,39.833]],[\"comment/252\",[]],[\"name/253\",[31,38.402]],[\"comment/253\",[]],[\"name/254\",[32,38.402]],[\"comment/254\",[]],[\"name/255\",[79,54.496]],[\"comment/255\",[]],[\"name/256\",[80,54.496]],[\"comment/256\",[]],[\"name/257\",[81,54.496]],[\"comment/257\",[]],[\"name/258\",[61,41.503]],[\"comment/258\",[]],[\"name/259\",[46,39.833]],[\"comment/259\",[]],[\"name/260\",[47,39.833]],[\"comment/260\",[]],[\"name/261\",[48,39.833]],[\"comment/261\",[]],[\"name/262\",[49,39.833]],[\"comment/262\",[]],[\"name/263\",[33,38.402]],[\"comment/263\",[]],[\"name/264\",[62,41.503]],[\"comment/264\",[]],[\"name/265\",[18,37.15]],[\"comment/265\",[]],[\"name/266\",[14,37.15]],[\"comment/266\",[]],[\"name/267\",[15,37.15]],[\"comment/267\",[]],[\"name/268\",[35,37.15]],[\"comment/268\",[]],[\"name/269\",[63,41.503]],[\"comment/269\",[]],[\"name/270\",[64,41.503]],[\"comment/270\",[]],[\"name/271\",[65,41.503]],[\"comment/271\",[]],[\"name/272\",[16,37.15]],[\"comment/272\",[]],[\"name/273\",[22,37.15]],[\"comment/273\",[]],[\"name/274\",[23,37.15]],[\"comment/274\",[]],[\"name/275\",[36,38.402]],[\"comment/275\",[]],[\"name/276\",[37,38.402]],[\"comment/276\",[]],[\"name/277\",[52,39.833]],[\"comment/277\",[]],[\"name/278\",[38,38.402]],[\"comment/278\",[]],[\"name/279\",[39,38.402]],[\"comment/279\",[]],[\"name/280\",[43,38.402]],[\"comment/280\",[]],[\"name/281\",[42,37.15]],[\"comment/281\",[]],[\"name/282\",[66,41.503]],[\"comment/282\",[]],[\"name/283\",[67,41.503]],[\"comment/283\",[]],[\"name/284\",[68,41.503]],[\"comment/284\",[]],[\"name/285\",[69,41.503]],[\"comment/285\",[]],[\"name/286\",[70,39.833]],[\"comment/286\",[]],[\"name/287\",[71,41.503]],[\"comment/287\",[]],[\"name/288\",[72,41.503]],[\"comment/288\",[]],[\"name/289\",[73,41.503]],[\"comment/289\",[]],[\"name/290\",[74,39.833]],[\"comment/290\",[]],[\"name/291\",[35,37.15]],[\"comment/291\",[]],[\"name/292\",[82,54.496]],[\"comment/292\",[]],[\"name/293\",[83,54.496]],[\"comment/293\",[]],[\"name/294\",[84,54.496]],[\"comment/294\",[]],[\"name/295\",[85,54.496]],[\"comment/295\",[]],[\"name/296\",[74,39.833]],[\"comment/296\",[]],[\"name/297\",[86,54.496]],[\"comment/297\",[]],[\"name/298\",[87,54.496]],[\"comment/298\",[]],[\"name/299\",[88,54.496]],[\"comment/299\",[]],[\"name/300\",[89,54.496]],[\"comment/300\",[]],[\"name/301\",[90,54.496]],[\"comment/301\",[]],[\"name/302\",[91,54.496]],[\"comment/302\",[]],[\"name/303\",[92,54.496]],[\"comment/303\",[]],[\"name/304\",[93,54.496]],[\"comment/304\",[]],[\"name/305\",[12,46.023]],[\"comment/305\",[]],[\"name/306\",[42,37.15]],[\"comment/306\",[]],[\"name/307\",[94,54.496]],[\"comment/307\",[]],[\"name/308\",[95,54.496]],[\"comment/308\",[]],[\"name/309\",[96,54.496]],[\"comment/309\",[]],[\"name/310\",[97,54.496]],[\"comment/310\",[]],[\"name/311\",[98,49.388]],[\"comment/311\",[]],[\"name/312\",[99,49.388]],[\"comment/312\",[]],[\"name/313\",[100,54.496]],[\"comment/313\",[]],[\"name/314\",[101,49.388]],[\"comment/314\",[]],[\"name/315\",[102,49.388]],[\"comment/315\",[]],[\"name/316\",[99,49.388]],[\"comment/316\",[]],[\"name/317\",[98,49.388]],[\"comment/317\",[]],[\"name/318\",[103,54.496]],[\"comment/318\",[]],[\"name/319\",[104,54.496]],[\"comment/319\",[]],[\"name/320\",[105,54.496]],[\"comment/320\",[]],[\"name/321\",[106,54.496]],[\"comment/321\",[]],[\"name/322\",[107,54.496]],[\"comment/322\",[]],[\"name/323\",[108,54.496]],[\"comment/323\",[]],[\"name/324\",[109,54.496]],[\"comment/324\",[]],[\"name/325\",[110,54.496]],[\"comment/325\",[]],[\"name/326\",[111,54.496]],[\"comment/326\",[]],[\"name/327\",[112,54.496]],[\"comment/327\",[]],[\"name/328\",[113,54.496]],[\"comment/328\",[]],[\"name/329\",[114,54.496]],[\"comment/329\",[]],[\"name/330\",[115,54.496]],[\"comment/330\",[]],[\"name/331\",[116,54.496]],[\"comment/331\",[]],[\"name/332\",[117,54.496]],[\"comment/332\",[]],[\"name/333\",[118,54.496]],[\"comment/333\",[]],[\"name/334\",[119,54.496]],[\"comment/334\",[]],[\"name/335\",[101,49.388]],[\"comment/335\",[]],[\"name/336\",[102,49.388]],[\"comment/336\",[]],[\"name/337\",[120,54.496]],[\"comment/337\",[]],[\"name/338\",[121,54.496]],[\"comment/338\",[]],[\"name/339\",[122,54.496]],[\"comment/339\",[]],[\"name/340\",[123,54.496]],[\"comment/340\",[]],[\"name/341\",[124,54.496]],[\"comment/341\",[]],[\"name/342\",[125,54.496]],[\"comment/342\",[]],[\"name/343\",[126,54.496]],[\"comment/343\",[]],[\"name/344\",[127,54.496]],[\"comment/344\",[]],[\"name/345\",[128,54.496]],[\"comment/345\",[]],[\"name/346\",[129,54.496]],[\"comment/346\",[]],[\"name/347\",[130,54.496]],[\"comment/347\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":26,\"name\":{\"31\":{},\"56\":{}},\"comment\":{}}],[\"_height\",{\"_index\":76,\"name\":{\"174\":{},\"215\":{}},\"comment\":{}}],[\"_radiusx\",{\"_index\":54,\"name\":{\"91\":{}},\"comment\":{}}],[\"_radiusy\",{\"_index\":55,\"name\":{\"92\":{}},\"comment\":{}}],[\"_step\",{\"_index\":56,\"name\":{\"93\":{}},\"comment\":{}}],[\"_width\",{\"_index\":75,\"name\":{\"173\":{},\"214\":{}},\"comment\":{}}],[\"angle\",{\"_index\":17,\"name\":{\"18\":{},\"53\":{},\"67\":{}},\"comment\":{}}],[\"bbox\",{\"_index\":33,\"name\":{\"39\":{},\"63\":{},\"104\":{},\"138\":{},\"182\":{},\"223\":{},\"263\":{}},\"comment\":{}}],[\"body\",{\"_index\":20,\"name\":{\"22\":{},\"35\":{}},\"comment\":{}}],[\"bodymoved\",{\"_index\":108,\"name\":{\"323\":{}},\"comment\":{}}],[\"bodyoptions\",{\"_index\":13,\"name\":{\"14\":{}},\"comment\":{}}],[\"bodyprops\",{\"_index\":30,\"name\":{\"36\":{}},\"comment\":{}}],[\"bodytype\",{\"_index\":2,\"name\":{\"2\":{}},\"comment\":{}}],[\"box\",{\"_index\":6,\"name\":{\"6\":{},\"169\":{}},\"comment\":{}}],[\"center\",{\"_index\":60,\"name\":{\"98\":{}},\"comment\":{}}],[\"centered\",{\"_index\":64,\"name\":{\"111\":{},\"147\":{},\"189\":{},\"230\":{},\"270\":{}},\"comment\":{}}],[\"checkainb\",{\"_index\":111,\"name\":{\"326\":{}},\"comment\":{}}],[\"checkall\",{\"_index\":89,\"name\":{\"300\":{}},\"comment\":{}}],[\"checkcollision\",{\"_index\":91,\"name\":{\"302\":{}},\"comment\":{}}],[\"checkone\",{\"_index\":88,\"name\":{\"299\":{}},\"comment\":{}}],[\"children\",{\"_index\":11,\"name\":{\"11\":{}},\"comment\":{}}],[\"childrendata\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"circle\",{\"_index\":5,\"name\":{\"5\":{},\"57\":{}},\"comment\":{}}],[\"circleincircle\",{\"_index\":124,\"name\":{\"341\":{}},\"comment\":{}}],[\"circleinpolygon\",{\"_index\":125,\"name\":{\"342\":{}},\"comment\":{}}],[\"circleoutsidepolygon\",{\"_index\":126,\"name\":{\"343\":{}},\"comment\":{}}],[\"clockwise\",{\"_index\":106,\"name\":{\"321\":{}},\"comment\":{}}],[\"clonepointsarray\",{\"_index\":112,\"name\":{\"327\":{}},\"comment\":{}}],[\"constructor\",{\"_index\":45,\"name\":{\"58\":{},\"88\":{},\"133\":{},\"170\":{},\"211\":{},\"252\":{}},\"comment\":{}}],[\"convexpolygons\",{\"_index\":62,\"name\":{\"105\":{},\"140\":{},\"183\":{},\"224\":{},\"264\":{}},\"comment\":{}}],[\"createbox\",{\"_index\":98,\"name\":{\"311\":{},\"317\":{}},\"comment\":{}}],[\"createcircle\",{\"_index\":97,\"name\":{\"310\":{}},\"comment\":{}}],[\"createellipse\",{\"_index\":99,\"name\":{\"312\":{},\"316\":{}},\"comment\":{}}],[\"createline\",{\"_index\":96,\"name\":{\"309\":{}},\"comment\":{}}],[\"createpoint\",{\"_index\":95,\"name\":{\"308\":{}},\"comment\":{}}],[\"createpolygon\",{\"_index\":100,\"name\":{\"313\":{}},\"comment\":{}}],[\"dashlineto\",{\"_index\":113,\"name\":{\"328\":{}},\"comment\":{}}],[\"data\",{\"_index\":12,\"name\":{\"12\":{},\"13\":{},\"305\":{}},\"comment\":{}}],[\"decomppoint\",{\"_index\":1,\"name\":{\"1\":{}},\"comment\":{}}],[\"decomppolygon\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"deg2rad\",{\"_index\":101,\"name\":{\"314\":{},\"335\":{}},\"comment\":{}}],[\"distance\",{\"_index\":105,\"name\":{\"320\":{}},\"comment\":{}}],[\"draw\",{\"_index\":42,\"name\":{\"48\":{},\"85\":{},\"121\":{},\"158\":{},\"200\":{},\"241\":{},\"281\":{},\"306\":{}},\"comment\":{}}],[\"drawbvh\",{\"_index\":94,\"name\":{\"307\":{}},\"comment\":{}}],[\"drawpolygon\",{\"_index\":119,\"name\":{\"334\":{}},\"comment\":{}}],[\"ellipse\",{\"_index\":3,\"name\":{\"3\":{},\"87\":{}},\"comment\":{}}],[\"end\",{\"_index\":80,\"name\":{\"256\":{}},\"comment\":{}}],[\"ensureconvex\",{\"_index\":116,\"name\":{\"331\":{}},\"comment\":{}}],[\"ensurepolygonpoints\",{\"_index\":104,\"name\":{\"319\":{}},\"comment\":{}}],[\"ensurevectorpoint\",{\"_index\":103,\"name\":{\"318\":{}},\"comment\":{}}],[\"extendbody\",{\"_index\":107,\"name\":{\"322\":{}},\"comment\":{}}],[\"getaabbasbbox\",{\"_index\":43,\"name\":{\"49\":{},\"84\":{},\"120\":{},\"157\":{},\"199\":{},\"240\":{},\"280\":{}},\"comment\":{}}],[\"getaabbasbox\",{\"_index\":25,\"name\":{\"29\":{},\"30\":{}},\"comment\":{}}],[\"getbouncedirection\",{\"_index\":117,\"name\":{\"332\":{}},\"comment\":{}}],[\"getcentroid\",{\"_index\":81,\"name\":{\"257\":{}},\"comment\":{}}],[\"getcentroidwithoutrotation\",{\"_index\":66,\"name\":{\"122\":{},\"159\":{},\"201\":{},\"242\":{},\"282\":{}},\"comment\":{}}],[\"getconvex\",{\"_index\":72,\"name\":{\"128\":{},\"165\":{},\"207\":{},\"248\":{},\"288\":{}},\"comment\":{}}],[\"getoffsetwithangle\",{\"_index\":53,\"name\":{\"86\":{}},\"comment\":{}}],[\"getpotentials\",{\"_index\":90,\"name\":{\"301\":{}},\"comment\":{}}],[\"getsattest\",{\"_index\":118,\"name\":{\"333\":{}},\"comment\":{}}],[\"h\",{\"_index\":29,\"name\":{\"34\":{}},\"comment\":{}}],[\"height\",{\"_index\":78,\"name\":{\"176\":{},\"217\":{}},\"comment\":{}}],[\"insert\",{\"_index\":85,\"name\":{\"295\":{}},\"comment\":{}}],[\"intersectaabb\",{\"_index\":110,\"name\":{\"325\":{}},\"comment\":{}}],[\"intersectlinecircle\",{\"_index\":127,\"name\":{\"344\":{}},\"comment\":{}}],[\"intersectlineline\",{\"_index\":129,\"name\":{\"346\":{}},\"comment\":{}}],[\"intersectlinelinefast\",{\"_index\":128,\"name\":{\"345\":{}},\"comment\":{}}],[\"intersectlinepolygon\",{\"_index\":130,\"name\":{\"347\":{}},\"comment\":{}}],[\"iscentered\",{\"_index\":16,\"name\":{\"17\":{},\"52\":{},\"73\":{},\"94\":{},\"149\":{},\"191\":{},\"232\":{},\"272\":{}},\"comment\":{}}],[\"isconvex\",{\"_index\":32,\"name\":{\"38\":{},\"71\":{},\"90\":{},\"139\":{},\"172\":{},\"213\":{},\"254\":{}},\"comment\":{}}],[\"issimple\",{\"_index\":70,\"name\":{\"126\":{},\"131\":{},\"163\":{},\"205\":{},\"246\":{},\"286\":{}},\"comment\":{}}],[\"isstatic\",{\"_index\":14,\"name\":{\"15\":{},\"50\":{},\"68\":{},\"107\":{},\"142\":{},\"185\":{},\"226\":{},\"266\":{}},\"comment\":{}}],[\"istrigger\",{\"_index\":15,\"name\":{\"16\":{},\"51\":{},\"69\":{},\"108\":{},\"143\":{},\"186\":{},\"227\":{},\"267\":{}},\"comment\":{}}],[\"leaf\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"line\",{\"_index\":4,\"name\":{\"4\":{},\"251\":{}},\"comment\":{}}],[\"maparraytovector\",{\"_index\":115,\"name\":{\"330\":{}},\"comment\":{}}],[\"mapvectortoarray\",{\"_index\":114,\"name\":{\"329\":{}},\"comment\":{}}],[\"maxx\",{\"_index\":47,\"name\":{\"60\":{},\"101\":{},\"135\":{},\"179\":{},\"220\":{},\"260\":{}},\"comment\":{}}],[\"maxy\",{\"_index\":49,\"name\":{\"62\":{},\"103\":{},\"137\":{},\"181\":{},\"222\":{},\"262\":{}},\"comment\":{}}],[\"minx\",{\"_index\":46,\"name\":{\"59\":{},\"100\":{},\"134\":{},\"178\":{},\"219\":{},\"259\":{}},\"comment\":{}}],[\"miny\",{\"_index\":48,\"name\":{\"61\":{},\"102\":{},\"136\":{},\"180\":{},\"221\":{},\"261\":{}},\"comment\":{}}],[\"notintersectaabb\",{\"_index\":109,\"name\":{\"324\":{}},\"comment\":{}}],[\"offset\",{\"_index\":34,\"name\":{\"40\":{},\"64\":{}},\"comment\":{}}],[\"offsetcopy\",{\"_index\":50,\"name\":{\"65\":{}},\"comment\":{}}],[\"padding\",{\"_index\":18,\"name\":{\"19\":{},\"54\":{},\"66\":{},\"106\":{},\"141\":{},\"184\":{},\"225\":{},\"265\":{}},\"comment\":{}}],[\"point\",{\"_index\":7,\"name\":{\"7\":{},\"21\":{},\"210\":{}},\"comment\":{}}],[\"pointinpolygon\",{\"_index\":121,\"name\":{\"338\":{}},\"comment\":{}}],[\"pointoncircle\",{\"_index\":123,\"name\":{\"340\":{}},\"comment\":{}}],[\"pointsbackup\",{\"_index\":63,\"name\":{\"110\":{},\"146\":{},\"188\":{},\"229\":{},\"269\":{}},\"comment\":{}}],[\"polygon\",{\"_index\":8,\"name\":{\"8\":{},\"132\":{}},\"comment\":{}}],[\"polygonincircle\",{\"_index\":120,\"name\":{\"337\":{}},\"comment\":{}}],[\"polygoninpolygon\",{\"_index\":122,\"name\":{\"339\":{}},\"comment\":{}}],[\"pos\",{\"_index\":27,\"name\":{\"32\":{}},\"comment\":{}}],[\"potentialvector\",{\"_index\":21,\"name\":{\"23\":{}},\"comment\":{}}],[\"rad2deg\",{\"_index\":102,\"name\":{\"315\":{},\"336\":{}},\"comment\":{}}],[\"radiusx\",{\"_index\":58,\"name\":{\"96\":{}},\"comment\":{}}],[\"radiusy\",{\"_index\":59,\"name\":{\"97\":{}},\"comment\":{}}],[\"ray\",{\"_index\":83,\"name\":{\"293\":{}},\"comment\":{}}],[\"raycast\",{\"_index\":92,\"name\":{\"303\":{}},\"comment\":{}}],[\"raycasthit\",{\"_index\":19,\"name\":{\"20\":{}},\"comment\":{}}],[\"remove\",{\"_index\":84,\"name\":{\"294\":{}},\"comment\":{}}],[\"response\",{\"_index\":82,\"name\":{\"292\":{}},\"comment\":{}}],[\"rotate\",{\"_index\":69,\"name\":{\"125\":{},\"162\":{},\"204\":{},\"245\":{},\"285\":{}},\"comment\":{}}],[\"sattest\",{\"_index\":44,\"name\":{\"55\":{}},\"comment\":{}}],[\"scale\",{\"_index\":52,\"name\":{\"77\":{},\"117\":{},\"154\":{},\"196\":{},\"237\":{},\"277\":{}},\"comment\":{}}],[\"scalevector\",{\"_index\":65,\"name\":{\"112\":{},\"148\":{},\"190\":{},\"231\":{},\"271\":{}},\"comment\":{}}],[\"scalex\",{\"_index\":36,\"name\":{\"42\":{},\"78\":{},\"115\":{},\"152\":{},\"194\":{},\"235\":{},\"275\":{}},\"comment\":{}}],[\"scaley\",{\"_index\":37,\"name\":{\"43\":{},\"79\":{},\"116\":{},\"153\":{},\"195\":{},\"236\":{},\"276\":{}},\"comment\":{}}],[\"separate\",{\"_index\":87,\"name\":{\"298\":{}},\"comment\":{}}],[\"setangle\",{\"_index\":40,\"name\":{\"46\":{},\"82\":{}},\"comment\":{}}],[\"setoffset\",{\"_index\":41,\"name\":{\"47\":{},\"83\":{}},\"comment\":{}}],[\"setpoints\",{\"_index\":67,\"name\":{\"123\":{},\"160\":{},\"202\":{},\"243\":{},\"283\":{}},\"comment\":{}}],[\"setposition\",{\"_index\":38,\"name\":{\"44\":{},\"80\":{},\"118\":{},\"155\":{},\"197\":{},\"238\":{},\"278\":{}},\"comment\":{}}],[\"setscale\",{\"_index\":39,\"name\":{\"45\":{},\"81\":{},\"119\":{},\"156\":{},\"198\":{},\"239\":{},\"279\":{}},\"comment\":{}}],[\"start\",{\"_index\":79,\"name\":{\"255\":{}},\"comment\":{}}],[\"step\",{\"_index\":57,\"name\":{\"95\":{}},\"comment\":{}}],[\"system\",{\"_index\":35,\"name\":{\"41\":{},\"70\":{},\"109\":{},\"144\":{},\"187\":{},\"228\":{},\"268\":{},\"291\":{}},\"comment\":{}}],[\"translate\",{\"_index\":68,\"name\":{\"124\":{},\"161\":{},\"203\":{},\"244\":{},\"284\":{}},\"comment\":{}}],[\"traverse\",{\"_index\":93,\"name\":{\"304\":{}},\"comment\":{}}],[\"type\",{\"_index\":31,\"name\":{\"37\":{},\"72\":{},\"89\":{},\"145\":{},\"171\":{},\"212\":{},\"253\":{}},\"comment\":{}}],[\"unscaledradius\",{\"_index\":51,\"name\":{\"74\":{}},\"comment\":{}}],[\"update\",{\"_index\":86,\"name\":{\"297\":{}},\"comment\":{}}],[\"updatebody\",{\"_index\":74,\"name\":{\"130\":{},\"168\":{},\"209\":{},\"250\":{},\"290\":{},\"296\":{}},\"comment\":{}}],[\"updateconvexpolygonpositions\",{\"_index\":71,\"name\":{\"127\":{},\"164\":{},\"206\":{},\"247\":{},\"287\":{}},\"comment\":{}}],[\"updateconvexpolygons\",{\"_index\":73,\"name\":{\"129\":{},\"166\":{},\"208\":{},\"249\":{},\"289\":{}},\"comment\":{}}],[\"updateisconvex\",{\"_index\":61,\"name\":{\"99\":{},\"167\":{},\"177\":{},\"218\":{},\"258\":{}},\"comment\":{}}],[\"vector\",{\"_index\":24,\"name\":{\"26\":{}},\"comment\":{}}],[\"w\",{\"_index\":28,\"name\":{\"33\":{}},\"comment\":{}}],[\"width\",{\"_index\":77,\"name\":{\"175\":{},\"216\":{}},\"comment\":{}}],[\"x\",{\"_index\":22,\"name\":{\"24\":{},\"27\":{},\"75\":{},\"113\":{},\"150\":{},\"192\":{},\"233\":{},\"273\":{}},\"comment\":{}}],[\"y\",{\"_index\":23,\"name\":{\"25\":{},\"28\":{},\"76\":{},\"114\":{},\"151\":{},\"193\":{},\"234\":{},\"274\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
+window.searchData = JSON.parse("{\"kinds\":{\"8\":\"Enumeration\",\"16\":\"Enumeration Member\",\"32\":\"Variable\",\"64\":\"Function\",\"128\":\"Class\",\"256\":\"Interface\",\"512\":\"Constructor\",\"1024\":\"Property\",\"2048\":\"Method\",\"65536\":\"Type literal\",\"262144\":\"Accessor\",\"4194304\":\"Type alias\"},\"rows\":[{\"kind\":4194304,\"name\":\"DecompPolygon\",\"url\":\"types/DecompPolygon.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"DecompPoint\",\"url\":\"types/DecompPoint.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":8,\"name\":\"BodyType\",\"url\":\"enums/BodyType.html\",\"classes\":\"tsd-kind-enum\"},{\"kind\":16,\"name\":\"Ellipse\",\"url\":\"enums/BodyType.html#Ellipse\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Line\",\"url\":\"enums/BodyType.html#Line\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Circle\",\"url\":\"enums/BodyType.html#Circle\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Box\",\"url\":\"enums/BodyType.html#Box\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Point\",\"url\":\"enums/BodyType.html#Point\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":16,\"name\":\"Polygon\",\"url\":\"enums/BodyType.html#Polygon\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"BodyType\"},{\"kind\":4194304,\"name\":\"Leaf\",\"url\":\"types/Leaf.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":256,\"name\":\"ChildrenData\",\"url\":\"interfaces/ChildrenData.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"children\",\"url\":\"interfaces/ChildrenData.html#children\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ChildrenData\"},{\"kind\":256,\"name\":\"Data\",\"url\":\"interfaces/Data.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"data\",\"url\":\"interfaces/Data.html#data\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Data\"},{\"kind\":256,\"name\":\"BodyOptions\",\"url\":\"interfaces/BodyOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"interfaces/BodyOptions.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"interfaces/BodyOptions.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":1024,\"name\":\"isCentered\",\"url\":\"interfaces/BodyOptions.html#isCentered\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"interfaces/BodyOptions.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"interfaces/BodyOptions.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyOptions\"},{\"kind\":256,\"name\":\"RaycastHit\",\"url\":\"interfaces/RaycastHit.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"point\",\"url\":\"interfaces/RaycastHit.html#point\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"RaycastHit\"},{\"kind\":1024,\"name\":\"body\",\"url\":\"interfaces/RaycastHit.html#body\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"RaycastHit\"},{\"kind\":256,\"name\":\"PotentialVector\",\"url\":\"interfaces/PotentialVector.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"x\",\"url\":\"interfaces/PotentialVector.html#x\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"PotentialVector\"},{\"kind\":1024,\"name\":\"y\",\"url\":\"interfaces/PotentialVector.html#y\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"PotentialVector\"},{\"kind\":256,\"name\":\"Vector\",\"url\":\"interfaces/Vector.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"x\",\"url\":\"interfaces/Vector.html#x\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Vector\"},{\"kind\":1024,\"name\":\"y\",\"url\":\"interfaces/Vector.html#y\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Vector\"},{\"kind\":256,\"name\":\"GetAABBAsBox\",\"url\":\"interfaces/GetAABBAsBox.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":2048,\"name\":\"getAABBAsBox\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"GetAABBAsBox\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox.getAABBAsBox-1.__type\",\"classes\":\"tsd-kind-type-literal\",\"parent\":\"GetAABBAsBox.getAABBAsBox.getAABBAsBox\"},{\"kind\":1024,\"name\":\"pos\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox.getAABBAsBox-1.__type.pos\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"GetAABBAsBox.getAABBAsBox.getAABBAsBox.__type\"},{\"kind\":1024,\"name\":\"w\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox.getAABBAsBox-1.__type.w\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"GetAABBAsBox.getAABBAsBox.getAABBAsBox.__type\"},{\"kind\":1024,\"name\":\"h\",\"url\":\"interfaces/GetAABBAsBox.html#getAABBAsBox.getAABBAsBox-1.__type.h\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"GetAABBAsBox.getAABBAsBox.getAABBAsBox.__type\"},{\"kind\":4194304,\"name\":\"Body\",\"url\":\"types/Body.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":256,\"name\":\"BodyProps\",\"url\":\"interfaces/BodyProps.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"interfaces/BodyProps.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"interfaces/BodyProps.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"interfaces/BodyProps.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"offset\",\"url\":\"interfaces/BodyProps.html#offset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"interfaces/BodyProps.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"interfaces/BodyProps.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"interfaces/BodyProps.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"interfaces/BodyProps.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"interfaces/BodyProps.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"setAngle\",\"url\":\"interfaces/BodyProps.html#setAngle\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"setOffset\",\"url\":\"interfaces/BodyProps.html#setOffset\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"interfaces/BodyProps.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"interfaces/BodyProps.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-interface\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"interfaces/BodyProps.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"interfaces/BodyProps.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"isCentered\",\"url\":\"interfaces/BodyProps.html#isCentered\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"interfaces/BodyProps.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"interfaces/BodyProps.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"BodyProps\"},{\"kind\":4194304,\"name\":\"SATTest\",\"url\":\"types/SATTest.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/SATTest.html#__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-type-alias\",\"parent\":\"SATTest\"},{\"kind\":128,\"name\":\"Circle\",\"url\":\"classes/Circle.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Circle.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Circle.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Circle.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Circle.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Circle.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Circle.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"offset\",\"url\":\"classes/Circle.html#offset\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"offsetCopy\",\"url\":\"classes/Circle.html#offsetCopy\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Circle.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"classes/Circle.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Circle.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Circle.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Circle.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Circle.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Circle.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"isCentered\",\"url\":\"classes/Circle.html#isCentered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":1024,\"name\":\"unscaledRadius\",\"url\":\"classes/Circle.html#unscaledRadius\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Circle.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Circle.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Circle.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Circle.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Circle.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Circle.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Circle.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"setAngle\",\"url\":\"classes/Circle.html#setAngle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"setOffset\",\"url\":\"classes/Circle.html#setOffset\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Circle.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Circle.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circle\"},{\"kind\":2048,\"name\":\"getOffsetWithAngle\",\"url\":\"classes/Circle.html#getOffsetWithAngle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Circle\"},{\"kind\":128,\"name\":\"Ellipse\",\"url\":\"classes/Ellipse.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Ellipse.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Ellipse.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Ellipse.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"_radiusX\",\"url\":\"classes/Ellipse.html#_radiusX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"_radiusY\",\"url\":\"classes/Ellipse.html#_radiusY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"_step\",\"url\":\"classes/Ellipse.html#_step\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Ellipse.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"step\",\"url\":\"classes/Ellipse.html#step\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"radiusX\",\"url\":\"classes/Ellipse.html#radiusX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"radiusY\",\"url\":\"classes/Ellipse.html#radiusY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"center\",\"url\":\"classes/Ellipse.html#center\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Ellipse.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Ellipse.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Ellipse.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Ellipse.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Ellipse.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Ellipse.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Ellipse.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Ellipse.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Ellipse.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Ellipse.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Ellipse.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Ellipse.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Ellipse.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Ellipse.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Ellipse.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Ellipse.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Ellipse.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Ellipse.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Ellipse.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Ellipse.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Ellipse.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Ellipse.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Ellipse.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Ellipse.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Ellipse.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Ellipse.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Ellipse.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Ellipse.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Ellipse.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Ellipse.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Ellipse.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Ellipse.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Ellipse\"},{\"kind\":64,\"name\":\"isSimple\",\"url\":\"functions/isSimple.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":128,\"name\":\"Polygon\",\"url\":\"classes/Polygon.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Polygon.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Polygon.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Polygon.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Polygon.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Polygon.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Polygon.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Polygon.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Polygon.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Polygon.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Polygon.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Polygon.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Polygon.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Polygon.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Polygon.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Polygon.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Polygon.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Polygon.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Polygon.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Polygon.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Polygon.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Polygon.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Polygon.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Polygon.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Polygon.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Polygon.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Polygon.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Polygon.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Polygon.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Polygon.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Polygon.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Polygon.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Polygon.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Polygon.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Polygon.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Polygon.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Polygon.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Polygon\"},{\"kind\":128,\"name\":\"Box\",\"url\":\"classes/Box.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Box.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Box.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Box.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"_width\",\"url\":\"classes/Box.html#_width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"_height\",\"url\":\"classes/Box.html#_height\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"width\",\"url\":\"classes/Box.html#width\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"height\",\"url\":\"classes/Box.html#height\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Box.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Box.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Box.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Box.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Box.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Box.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Box.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Box.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Box.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Box.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Box.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Box.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Box.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Box.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Box.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Box.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Box.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Box.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Box.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Box.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Box.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Box.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Box.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Box.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Box.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Box.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Box.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Box.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Box.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Box.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Box.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Box.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Box.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Box\"},{\"kind\":128,\"name\":\"Point\",\"url\":\"classes/Point.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Point.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Point.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Point.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"_width\",\"url\":\"classes/Point.html#_width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"_height\",\"url\":\"classes/Point.html#_height\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"width\",\"url\":\"classes/Point.html#width\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"height\",\"url\":\"classes/Point.html#height\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Point.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Point.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Point.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Point.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Point.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Point.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Point.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Point.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Point.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Point.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Point.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Point.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Point.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Point.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Point.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Point.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Point.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Point.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Point.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Point.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Point.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Point.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Point.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Point.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Point.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Point.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Point.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Point.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Point.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Point.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Point.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Point.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Point.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Point\"},{\"kind\":128,\"name\":\"Line\",\"url\":\"classes/Line.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Line.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"classes/Line.html#type\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"isConvex\",\"url\":\"classes/Line.html#isConvex\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"start\",\"url\":\"classes/Line.html#start\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"end\",\"url\":\"classes/Line.html#end\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"getCentroid\",\"url\":\"classes/Line.html#getCentroid\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"updateIsConvex\",\"url\":\"classes/Line.html#updateIsConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"minX\",\"url\":\"classes/Line.html#minX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"maxX\",\"url\":\"classes/Line.html#maxX\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"minY\",\"url\":\"classes/Line.html#minY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"maxY\",\"url\":\"classes/Line.html#maxY\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"bbox\",\"url\":\"classes/Line.html#bbox\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"convexPolygons\",\"url\":\"classes/Line.html#convexPolygons\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"padding\",\"url\":\"classes/Line.html#padding\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"isStatic\",\"url\":\"classes/Line.html#isStatic\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"isTrigger\",\"url\":\"classes/Line.html#isTrigger\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"system\",\"url\":\"classes/Line.html#system\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"pointsBackup\",\"url\":\"classes/Line.html#pointsBackup\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"centered\",\"url\":\"classes/Line.html#centered\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":1024,\"name\":\"scaleVector\",\"url\":\"classes/Line.html#scaleVector\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"isCentered\",\"url\":\"classes/Line.html#isCentered\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"x\",\"url\":\"classes/Line.html#x\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"y\",\"url\":\"classes/Line.html#y\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"scaleX\",\"url\":\"classes/Line.html#scaleX\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"scaleY\",\"url\":\"classes/Line.html#scaleY\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":262144,\"name\":\"scale\",\"url\":\"classes/Line.html#scale\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"setPosition\",\"url\":\"classes/Line.html#setPosition\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"setScale\",\"url\":\"classes/Line.html#setScale\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"getAABBAsBBox\",\"url\":\"classes/Line.html#getAABBAsBBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/Line.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"getCentroidWithoutRotation\",\"url\":\"classes/Line.html#getCentroidWithoutRotation\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"setPoints\",\"url\":\"classes/Line.html#setPoints\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"translate\",\"url\":\"classes/Line.html#translate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"rotate\",\"url\":\"classes/Line.html#rotate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"isSimple\",\"url\":\"classes/Line.html#isSimple\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"updateConvexPolygonPositions\",\"url\":\"classes/Line.html#updateConvexPolygonPositions\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"getConvex\",\"url\":\"classes/Line.html#getConvex\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"updateConvexPolygons\",\"url\":\"classes/Line.html#updateConvexPolygons\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/Line.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected tsd-is-inherited\",\"parent\":\"Line\"},{\"kind\":128,\"name\":\"System\",\"url\":\"classes/System.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":1024,\"name\":\"response\",\"url\":\"classes/System.html#response\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":1024,\"name\":\"ray\",\"url\":\"classes/System.html#ray\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-protected\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"remove\",\"url\":\"classes/System.html#remove\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"insert\",\"url\":\"classes/System.html#insert\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"updateBody\",\"url\":\"classes/System.html#updateBody\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"update\",\"url\":\"classes/System.html#update\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"separate\",\"url\":\"classes/System.html#separate\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"checkOne\",\"url\":\"classes/System.html#checkOne\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"checkAll\",\"url\":\"classes/System.html#checkAll\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"getPotentials\",\"url\":\"classes/System.html#getPotentials\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"checkCollision\",\"url\":\"classes/System.html#checkCollision\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"raycast\",\"url\":\"classes/System.html#raycast\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"traverse\",\"url\":\"classes/System.html#traverse\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"System\"},{\"kind\":1024,\"name\":\"data\",\"url\":\"classes/System.html#data\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"draw\",\"url\":\"classes/System.html#draw\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"drawBVH\",\"url\":\"classes/System.html#drawBVH\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createPoint\",\"url\":\"classes/System.html#createPoint\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createLine\",\"url\":\"classes/System.html#createLine\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createCircle\",\"url\":\"classes/System.html#createCircle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createBox\",\"url\":\"classes/System.html#createBox\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createEllipse\",\"url\":\"classes/System.html#createEllipse\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":2048,\"name\":\"createPolygon\",\"url\":\"classes/System.html#createPolygon\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"System\"},{\"kind\":64,\"name\":\"deg2rad\",\"url\":\"functions/deg2rad-1.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"rad2deg\",\"url\":\"functions/rad2deg-1.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"createEllipse\",\"url\":\"functions/createEllipse.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"createBox\",\"url\":\"functions/createBox.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"ensureVectorPoint\",\"url\":\"functions/ensureVectorPoint.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"ensurePolygonPoints\",\"url\":\"functions/ensurePolygonPoints.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"distance\",\"url\":\"functions/distance.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"clockwise\",\"url\":\"functions/clockwise.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"extendBody\",\"url\":\"functions/extendBody.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"bodyMoved\",\"url\":\"functions/bodyMoved.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"notIntersectAABB\",\"url\":\"functions/notIntersectAABB.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectAABB\",\"url\":\"functions/intersectAABB.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"checkAInB\",\"url\":\"functions/checkAInB.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"clonePointsArray\",\"url\":\"functions/clonePointsArray.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"dashLineTo\",\"url\":\"functions/dashLineTo.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"mapVectorToArray\",\"url\":\"functions/mapVectorToArray.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"mapArrayToVector\",\"url\":\"functions/mapArrayToVector.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"getBounceDirection\",\"url\":\"functions/getBounceDirection.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"getSATTest\",\"url\":\"functions/getSATTest.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"drawPolygon\",\"url\":\"functions/drawPolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":32,\"name\":\"DEG2RAD\",\"url\":\"variables/DEG2RAD.html\",\"classes\":\"tsd-kind-variable\"},{\"kind\":32,\"name\":\"RAD2DEG\",\"url\":\"variables/RAD2DEG.html\",\"classes\":\"tsd-kind-variable\"},{\"kind\":64,\"name\":\"ensureConvex\",\"url\":\"functions/ensureConvex.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"polygonInCircle\",\"url\":\"functions/polygonInCircle.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"pointInPolygon\",\"url\":\"functions/pointInPolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"polygonInPolygon\",\"url\":\"functions/polygonInPolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"pointOnCircle\",\"url\":\"functions/pointOnCircle.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"circleInCircle\",\"url\":\"functions/circleInCircle.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"circleInPolygon\",\"url\":\"functions/circleInPolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"circleOutsidePolygon\",\"url\":\"functions/circleOutsidePolygon.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectLineCircle\",\"url\":\"functions/intersectLineCircle.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectLineLineFast\",\"url\":\"functions/intersectLineLineFast.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectLineLine\",\"url\":\"functions/intersectLineLine.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"intersectLinePolygon\",\"url\":\"functions/intersectLinePolygon.html\",\"classes\":\"tsd-kind-function\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,54.496]],[\"comment/0\",[]],[\"name/1\",[1,54.496]],[\"comment/1\",[]],[\"name/2\",[2,54.496]],[\"comment/2\",[]],[\"name/3\",[3,49.388]],[\"comment/3\",[]],[\"name/4\",[4,49.388]],[\"comment/4\",[]],[\"name/5\",[5,49.388]],[\"comment/5\",[]],[\"name/6\",[6,49.388]],[\"comment/6\",[]],[\"name/7\",[7,46.023]],[\"comment/7\",[]],[\"name/8\",[8,49.388]],[\"comment/8\",[]],[\"name/9\",[9,54.496]],[\"comment/9\",[]],[\"name/10\",[10,54.496]],[\"comment/10\",[]],[\"name/11\",[11,54.496]],[\"comment/11\",[]],[\"name/12\",[12,46.023]],[\"comment/12\",[]],[\"name/13\",[12,46.023]],[\"comment/13\",[]],[\"name/14\",[13,54.496]],[\"comment/14\",[]],[\"name/15\",[14,37.15]],[\"comment/15\",[]],[\"name/16\",[15,37.15]],[\"comment/16\",[]],[\"name/17\",[16,37.15]],[\"comment/17\",[]],[\"name/18\",[17,46.023]],[\"comment/18\",[]],[\"name/19\",[18,37.15]],[\"comment/19\",[]],[\"name/20\",[19,54.496]],[\"comment/20\",[]],[\"name/21\",[7,46.023]],[\"comment/21\",[]],[\"name/22\",[20,49.388]],[\"comment/22\",[]],[\"name/23\",[21,54.496]],[\"comment/23\",[]],[\"name/24\",[22,37.15]],[\"comment/24\",[]],[\"name/25\",[23,37.15]],[\"comment/25\",[]],[\"name/26\",[24,54.496]],[\"comment/26\",[]],[\"name/27\",[22,37.15]],[\"comment/27\",[]],[\"name/28\",[23,37.15]],[\"comment/28\",[]],[\"name/29\",[25,49.388]],[\"comment/29\",[]],[\"name/30\",[25,49.388]],[\"comment/30\",[]],[\"name/31\",[26,49.388]],[\"comment/31\",[]],[\"name/32\",[27,54.496]],[\"comment/32\",[]],[\"name/33\",[28,54.496]],[\"comment/33\",[]],[\"name/34\",[29,54.496]],[\"comment/34\",[]],[\"name/35\",[20,49.388]],[\"comment/35\",[]],[\"name/36\",[30,54.496]],[\"comment/36\",[]],[\"name/37\",[31,38.402]],[\"comment/37\",[]],[\"name/38\",[32,38.402]],[\"comment/38\",[]],[\"name/39\",[33,38.402]],[\"comment/39\",[]],[\"name/40\",[34,49.388]],[\"comment/40\",[]],[\"name/41\",[35,37.15]],[\"comment/41\",[]],[\"name/42\",[36,38.402]],[\"comment/42\",[]],[\"name/43\",[37,38.402]],[\"comment/43\",[]],[\"name/44\",[38,38.402]],[\"comment/44\",[]],[\"name/45\",[39,38.402]],[\"comment/45\",[]],[\"name/46\",[40,49.388]],[\"comment/46\",[]],[\"name/47\",[41,49.388]],[\"comment/47\",[]],[\"name/48\",[42,37.15]],[\"comment/48\",[]],[\"name/49\",[43,38.402]],[\"comment/49\",[]],[\"name/50\",[14,37.15]],[\"comment/50\",[]],[\"name/51\",[15,37.15]],[\"comment/51\",[]],[\"name/52\",[16,37.15]],[\"comment/52\",[]],[\"name/53\",[17,46.023]],[\"comment/53\",[]],[\"name/54\",[18,37.15]],[\"comment/54\",[]],[\"name/55\",[44,54.496]],[\"comment/55\",[]],[\"name/56\",[26,49.388]],[\"comment/56\",[]],[\"name/57\",[5,49.388]],[\"comment/57\",[]],[\"name/58\",[45,39.833]],[\"comment/58\",[]],[\"name/59\",[46,39.833]],[\"comment/59\",[]],[\"name/60\",[47,39.833]],[\"comment/60\",[]],[\"name/61\",[48,39.833]],[\"comment/61\",[]],[\"name/62\",[49,39.833]],[\"comment/62\",[]],[\"name/63\",[33,38.402]],[\"comment/63\",[]],[\"name/64\",[34,49.388]],[\"comment/64\",[]],[\"name/65\",[50,54.496]],[\"comment/65\",[]],[\"name/66\",[18,37.15]],[\"comment/66\",[]],[\"name/67\",[17,46.023]],[\"comment/67\",[]],[\"name/68\",[14,37.15]],[\"comment/68\",[]],[\"name/69\",[15,37.15]],[\"comment/69\",[]],[\"name/70\",[35,37.15]],[\"comment/70\",[]],[\"name/71\",[32,38.402]],[\"comment/71\",[]],[\"name/72\",[31,38.402]],[\"comment/72\",[]],[\"name/73\",[16,37.15]],[\"comment/73\",[]],[\"name/74\",[51,54.496]],[\"comment/74\",[]],[\"name/75\",[22,37.15]],[\"comment/75\",[]],[\"name/76\",[23,37.15]],[\"comment/76\",[]],[\"name/77\",[52,39.833]],[\"comment/77\",[]],[\"name/78\",[36,38.402]],[\"comment/78\",[]],[\"name/79\",[37,38.402]],[\"comment/79\",[]],[\"name/80\",[38,38.402]],[\"comment/80\",[]],[\"name/81\",[39,38.402]],[\"comment/81\",[]],[\"name/82\",[40,49.388]],[\"comment/82\",[]],[\"name/83\",[41,49.388]],[\"comment/83\",[]],[\"name/84\",[43,38.402]],[\"comment/84\",[]],[\"name/85\",[42,37.15]],[\"comment/85\",[]],[\"name/86\",[53,54.496]],[\"comment/86\",[]],[\"name/87\",[3,49.388]],[\"comment/87\",[]],[\"name/88\",[45,39.833]],[\"comment/88\",[]],[\"name/89\",[31,38.402]],[\"comment/89\",[]],[\"name/90\",[32,38.402]],[\"comment/90\",[]],[\"name/91\",[54,54.496]],[\"comment/91\",[]],[\"name/92\",[55,54.496]],[\"comment/92\",[]],[\"name/93\",[56,54.496]],[\"comment/93\",[]],[\"name/94\",[16,37.15]],[\"comment/94\",[]],[\"name/95\",[57,54.496]],[\"comment/95\",[]],[\"name/96\",[58,54.496]],[\"comment/96\",[]],[\"name/97\",[59,54.496]],[\"comment/97\",[]],[\"name/98\",[60,54.496]],[\"comment/98\",[]],[\"name/99\",[61,41.503]],[\"comment/99\",[]],[\"name/100\",[46,39.833]],[\"comment/100\",[]],[\"name/101\",[47,39.833]],[\"comment/101\",[]],[\"name/102\",[48,39.833]],[\"comment/102\",[]],[\"name/103\",[49,39.833]],[\"comment/103\",[]],[\"name/104\",[33,38.402]],[\"comment/104\",[]],[\"name/105\",[62,41.503]],[\"comment/105\",[]],[\"name/106\",[18,37.15]],[\"comment/106\",[]],[\"name/107\",[14,37.15]],[\"comment/107\",[]],[\"name/108\",[15,37.15]],[\"comment/108\",[]],[\"name/109\",[35,37.15]],[\"comment/109\",[]],[\"name/110\",[63,41.503]],[\"comment/110\",[]],[\"name/111\",[64,41.503]],[\"comment/111\",[]],[\"name/112\",[65,41.503]],[\"comment/112\",[]],[\"name/113\",[22,37.15]],[\"comment/113\",[]],[\"name/114\",[23,37.15]],[\"comment/114\",[]],[\"name/115\",[36,38.402]],[\"comment/115\",[]],[\"name/116\",[37,38.402]],[\"comment/116\",[]],[\"name/117\",[52,39.833]],[\"comment/117\",[]],[\"name/118\",[38,38.402]],[\"comment/118\",[]],[\"name/119\",[39,38.402]],[\"comment/119\",[]],[\"name/120\",[43,38.402]],[\"comment/120\",[]],[\"name/121\",[42,37.15]],[\"comment/121\",[]],[\"name/122\",[66,41.503]],[\"comment/122\",[]],[\"name/123\",[67,41.503]],[\"comment/123\",[]],[\"name/124\",[68,41.503]],[\"comment/124\",[]],[\"name/125\",[69,41.503]],[\"comment/125\",[]],[\"name/126\",[70,39.833]],[\"comment/126\",[]],[\"name/127\",[71,41.503]],[\"comment/127\",[]],[\"name/128\",[72,41.503]],[\"comment/128\",[]],[\"name/129\",[73,41.503]],[\"comment/129\",[]],[\"name/130\",[74,39.833]],[\"comment/130\",[]],[\"name/131\",[70,39.833]],[\"comment/131\",[]],[\"name/132\",[8,49.388]],[\"comment/132\",[]],[\"name/133\",[45,39.833]],[\"comment/133\",[]],[\"name/134\",[46,39.833]],[\"comment/134\",[]],[\"name/135\",[47,39.833]],[\"comment/135\",[]],[\"name/136\",[48,39.833]],[\"comment/136\",[]],[\"name/137\",[49,39.833]],[\"comment/137\",[]],[\"name/138\",[33,38.402]],[\"comment/138\",[]],[\"name/139\",[32,38.402]],[\"comment/139\",[]],[\"name/140\",[62,41.503]],[\"comment/140\",[]],[\"name/141\",[18,37.15]],[\"comment/141\",[]],[\"name/142\",[14,37.15]],[\"comment/142\",[]],[\"name/143\",[15,37.15]],[\"comment/143\",[]],[\"name/144\",[35,37.15]],[\"comment/144\",[]],[\"name/145\",[31,38.402]],[\"comment/145\",[]],[\"name/146\",[63,41.503]],[\"comment/146\",[]],[\"name/147\",[64,41.503]],[\"comment/147\",[]],[\"name/148\",[65,41.503]],[\"comment/148\",[]],[\"name/149\",[16,37.15]],[\"comment/149\",[]],[\"name/150\",[22,37.15]],[\"comment/150\",[]],[\"name/151\",[23,37.15]],[\"comment/151\",[]],[\"name/152\",[36,38.402]],[\"comment/152\",[]],[\"name/153\",[37,38.402]],[\"comment/153\",[]],[\"name/154\",[52,39.833]],[\"comment/154\",[]],[\"name/155\",[38,38.402]],[\"comment/155\",[]],[\"name/156\",[39,38.402]],[\"comment/156\",[]],[\"name/157\",[43,38.402]],[\"comment/157\",[]],[\"name/158\",[42,37.15]],[\"comment/158\",[]],[\"name/159\",[66,41.503]],[\"comment/159\",[]],[\"name/160\",[67,41.503]],[\"comment/160\",[]],[\"name/161\",[68,41.503]],[\"comment/161\",[]],[\"name/162\",[69,41.503]],[\"comment/162\",[]],[\"name/163\",[70,39.833]],[\"comment/163\",[]],[\"name/164\",[71,41.503]],[\"comment/164\",[]],[\"name/165\",[72,41.503]],[\"comment/165\",[]],[\"name/166\",[73,41.503]],[\"comment/166\",[]],[\"name/167\",[61,41.503]],[\"comment/167\",[]],[\"name/168\",[74,39.833]],[\"comment/168\",[]],[\"name/169\",[6,49.388]],[\"comment/169\",[]],[\"name/170\",[45,39.833]],[\"comment/170\",[]],[\"name/171\",[31,38.402]],[\"comment/171\",[]],[\"name/172\",[32,38.402]],[\"comment/172\",[]],[\"name/173\",[75,49.388]],[\"comment/173\",[]],[\"name/174\",[76,49.388]],[\"comment/174\",[]],[\"name/175\",[77,49.388]],[\"comment/175\",[]],[\"name/176\",[78,49.388]],[\"comment/176\",[]],[\"name/177\",[61,41.503]],[\"comment/177\",[]],[\"name/178\",[46,39.833]],[\"comment/178\",[]],[\"name/179\",[47,39.833]],[\"comment/179\",[]],[\"name/180\",[48,39.833]],[\"comment/180\",[]],[\"name/181\",[49,39.833]],[\"comment/181\",[]],[\"name/182\",[33,38.402]],[\"comment/182\",[]],[\"name/183\",[62,41.503]],[\"comment/183\",[]],[\"name/184\",[18,37.15]],[\"comment/184\",[]],[\"name/185\",[14,37.15]],[\"comment/185\",[]],[\"name/186\",[15,37.15]],[\"comment/186\",[]],[\"name/187\",[35,37.15]],[\"comment/187\",[]],[\"name/188\",[63,41.503]],[\"comment/188\",[]],[\"name/189\",[64,41.503]],[\"comment/189\",[]],[\"name/190\",[65,41.503]],[\"comment/190\",[]],[\"name/191\",[16,37.15]],[\"comment/191\",[]],[\"name/192\",[22,37.15]],[\"comment/192\",[]],[\"name/193\",[23,37.15]],[\"comment/193\",[]],[\"name/194\",[36,38.402]],[\"comment/194\",[]],[\"name/195\",[37,38.402]],[\"comment/195\",[]],[\"name/196\",[52,39.833]],[\"comment/196\",[]],[\"name/197\",[38,38.402]],[\"comment/197\",[]],[\"name/198\",[39,38.402]],[\"comment/198\",[]],[\"name/199\",[43,38.402]],[\"comment/199\",[]],[\"name/200\",[42,37.15]],[\"comment/200\",[]],[\"name/201\",[66,41.503]],[\"comment/201\",[]],[\"name/202\",[67,41.503]],[\"comment/202\",[]],[\"name/203\",[68,41.503]],[\"comment/203\",[]],[\"name/204\",[69,41.503]],[\"comment/204\",[]],[\"name/205\",[70,39.833]],[\"comment/205\",[]],[\"name/206\",[71,41.503]],[\"comment/206\",[]],[\"name/207\",[72,41.503]],[\"comment/207\",[]],[\"name/208\",[73,41.503]],[\"comment/208\",[]],[\"name/209\",[74,39.833]],[\"comment/209\",[]],[\"name/210\",[7,46.023]],[\"comment/210\",[]],[\"name/211\",[45,39.833]],[\"comment/211\",[]],[\"name/212\",[31,38.402]],[\"comment/212\",[]],[\"name/213\",[32,38.402]],[\"comment/213\",[]],[\"name/214\",[75,49.388]],[\"comment/214\",[]],[\"name/215\",[76,49.388]],[\"comment/215\",[]],[\"name/216\",[77,49.388]],[\"comment/216\",[]],[\"name/217\",[78,49.388]],[\"comment/217\",[]],[\"name/218\",[61,41.503]],[\"comment/218\",[]],[\"name/219\",[46,39.833]],[\"comment/219\",[]],[\"name/220\",[47,39.833]],[\"comment/220\",[]],[\"name/221\",[48,39.833]],[\"comment/221\",[]],[\"name/222\",[49,39.833]],[\"comment/222\",[]],[\"name/223\",[33,38.402]],[\"comment/223\",[]],[\"name/224\",[62,41.503]],[\"comment/224\",[]],[\"name/225\",[18,37.15]],[\"comment/225\",[]],[\"name/226\",[14,37.15]],[\"comment/226\",[]],[\"name/227\",[15,37.15]],[\"comment/227\",[]],[\"name/228\",[35,37.15]],[\"comment/228\",[]],[\"name/229\",[63,41.503]],[\"comment/229\",[]],[\"name/230\",[64,41.503]],[\"comment/230\",[]],[\"name/231\",[65,41.503]],[\"comment/231\",[]],[\"name/232\",[16,37.15]],[\"comment/232\",[]],[\"name/233\",[22,37.15]],[\"comment/233\",[]],[\"name/234\",[23,37.15]],[\"comment/234\",[]],[\"name/235\",[36,38.402]],[\"comment/235\",[]],[\"name/236\",[37,38.402]],[\"comment/236\",[]],[\"name/237\",[52,39.833]],[\"comment/237\",[]],[\"name/238\",[38,38.402]],[\"comment/238\",[]],[\"name/239\",[39,38.402]],[\"comment/239\",[]],[\"name/240\",[43,38.402]],[\"comment/240\",[]],[\"name/241\",[42,37.15]],[\"comment/241\",[]],[\"name/242\",[66,41.503]],[\"comment/242\",[]],[\"name/243\",[67,41.503]],[\"comment/243\",[]],[\"name/244\",[68,41.503]],[\"comment/244\",[]],[\"name/245\",[69,41.503]],[\"comment/245\",[]],[\"name/246\",[70,39.833]],[\"comment/246\",[]],[\"name/247\",[71,41.503]],[\"comment/247\",[]],[\"name/248\",[72,41.503]],[\"comment/248\",[]],[\"name/249\",[73,41.503]],[\"comment/249\",[]],[\"name/250\",[74,39.833]],[\"comment/250\",[]],[\"name/251\",[4,49.388]],[\"comment/251\",[]],[\"name/252\",[45,39.833]],[\"comment/252\",[]],[\"name/253\",[31,38.402]],[\"comment/253\",[]],[\"name/254\",[32,38.402]],[\"comment/254\",[]],[\"name/255\",[79,54.496]],[\"comment/255\",[]],[\"name/256\",[80,54.496]],[\"comment/256\",[]],[\"name/257\",[81,54.496]],[\"comment/257\",[]],[\"name/258\",[61,41.503]],[\"comment/258\",[]],[\"name/259\",[46,39.833]],[\"comment/259\",[]],[\"name/260\",[47,39.833]],[\"comment/260\",[]],[\"name/261\",[48,39.833]],[\"comment/261\",[]],[\"name/262\",[49,39.833]],[\"comment/262\",[]],[\"name/263\",[33,38.402]],[\"comment/263\",[]],[\"name/264\",[62,41.503]],[\"comment/264\",[]],[\"name/265\",[18,37.15]],[\"comment/265\",[]],[\"name/266\",[14,37.15]],[\"comment/266\",[]],[\"name/267\",[15,37.15]],[\"comment/267\",[]],[\"name/268\",[35,37.15]],[\"comment/268\",[]],[\"name/269\",[63,41.503]],[\"comment/269\",[]],[\"name/270\",[64,41.503]],[\"comment/270\",[]],[\"name/271\",[65,41.503]],[\"comment/271\",[]],[\"name/272\",[16,37.15]],[\"comment/272\",[]],[\"name/273\",[22,37.15]],[\"comment/273\",[]],[\"name/274\",[23,37.15]],[\"comment/274\",[]],[\"name/275\",[36,38.402]],[\"comment/275\",[]],[\"name/276\",[37,38.402]],[\"comment/276\",[]],[\"name/277\",[52,39.833]],[\"comment/277\",[]],[\"name/278\",[38,38.402]],[\"comment/278\",[]],[\"name/279\",[39,38.402]],[\"comment/279\",[]],[\"name/280\",[43,38.402]],[\"comment/280\",[]],[\"name/281\",[42,37.15]],[\"comment/281\",[]],[\"name/282\",[66,41.503]],[\"comment/282\",[]],[\"name/283\",[67,41.503]],[\"comment/283\",[]],[\"name/284\",[68,41.503]],[\"comment/284\",[]],[\"name/285\",[69,41.503]],[\"comment/285\",[]],[\"name/286\",[70,39.833]],[\"comment/286\",[]],[\"name/287\",[71,41.503]],[\"comment/287\",[]],[\"name/288\",[72,41.503]],[\"comment/288\",[]],[\"name/289\",[73,41.503]],[\"comment/289\",[]],[\"name/290\",[74,39.833]],[\"comment/290\",[]],[\"name/291\",[35,37.15]],[\"comment/291\",[]],[\"name/292\",[82,54.496]],[\"comment/292\",[]],[\"name/293\",[83,54.496]],[\"comment/293\",[]],[\"name/294\",[84,54.496]],[\"comment/294\",[]],[\"name/295\",[85,54.496]],[\"comment/295\",[]],[\"name/296\",[74,39.833]],[\"comment/296\",[]],[\"name/297\",[86,54.496]],[\"comment/297\",[]],[\"name/298\",[87,54.496]],[\"comment/298\",[]],[\"name/299\",[88,54.496]],[\"comment/299\",[]],[\"name/300\",[89,54.496]],[\"comment/300\",[]],[\"name/301\",[90,54.496]],[\"comment/301\",[]],[\"name/302\",[91,54.496]],[\"comment/302\",[]],[\"name/303\",[92,54.496]],[\"comment/303\",[]],[\"name/304\",[93,54.496]],[\"comment/304\",[]],[\"name/305\",[12,46.023]],[\"comment/305\",[]],[\"name/306\",[42,37.15]],[\"comment/306\",[]],[\"name/307\",[94,54.496]],[\"comment/307\",[]],[\"name/308\",[95,54.496]],[\"comment/308\",[]],[\"name/309\",[96,54.496]],[\"comment/309\",[]],[\"name/310\",[97,54.496]],[\"comment/310\",[]],[\"name/311\",[98,49.388]],[\"comment/311\",[]],[\"name/312\",[99,49.388]],[\"comment/312\",[]],[\"name/313\",[100,54.496]],[\"comment/313\",[]],[\"name/314\",[101,49.388]],[\"comment/314\",[]],[\"name/315\",[102,49.388]],[\"comment/315\",[]],[\"name/316\",[99,49.388]],[\"comment/316\",[]],[\"name/317\",[98,49.388]],[\"comment/317\",[]],[\"name/318\",[103,54.496]],[\"comment/318\",[]],[\"name/319\",[104,54.496]],[\"comment/319\",[]],[\"name/320\",[105,54.496]],[\"comment/320\",[]],[\"name/321\",[106,54.496]],[\"comment/321\",[]],[\"name/322\",[107,54.496]],[\"comment/322\",[]],[\"name/323\",[108,54.496]],[\"comment/323\",[]],[\"name/324\",[109,54.496]],[\"comment/324\",[]],[\"name/325\",[110,54.496]],[\"comment/325\",[]],[\"name/326\",[111,54.496]],[\"comment/326\",[]],[\"name/327\",[112,54.496]],[\"comment/327\",[]],[\"name/328\",[113,54.496]],[\"comment/328\",[]],[\"name/329\",[114,54.496]],[\"comment/329\",[]],[\"name/330\",[115,54.496]],[\"comment/330\",[]],[\"name/331\",[116,54.496]],[\"comment/331\",[]],[\"name/332\",[117,54.496]],[\"comment/332\",[]],[\"name/333\",[118,54.496]],[\"comment/333\",[]],[\"name/334\",[101,49.388]],[\"comment/334\",[]],[\"name/335\",[102,49.388]],[\"comment/335\",[]],[\"name/336\",[119,54.496]],[\"comment/336\",[]],[\"name/337\",[120,54.496]],[\"comment/337\",[]],[\"name/338\",[121,54.496]],[\"comment/338\",[]],[\"name/339\",[122,54.496]],[\"comment/339\",[]],[\"name/340\",[123,54.496]],[\"comment/340\",[]],[\"name/341\",[124,54.496]],[\"comment/341\",[]],[\"name/342\",[125,54.496]],[\"comment/342\",[]],[\"name/343\",[126,54.496]],[\"comment/343\",[]],[\"name/344\",[127,54.496]],[\"comment/344\",[]],[\"name/345\",[128,54.496]],[\"comment/345\",[]],[\"name/346\",[129,54.496]],[\"comment/346\",[]],[\"name/347\",[130,54.496]],[\"comment/347\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":26,\"name\":{\"31\":{},\"56\":{}},\"comment\":{}}],[\"_height\",{\"_index\":76,\"name\":{\"174\":{},\"215\":{}},\"comment\":{}}],[\"_radiusx\",{\"_index\":54,\"name\":{\"91\":{}},\"comment\":{}}],[\"_radiusy\",{\"_index\":55,\"name\":{\"92\":{}},\"comment\":{}}],[\"_step\",{\"_index\":56,\"name\":{\"93\":{}},\"comment\":{}}],[\"_width\",{\"_index\":75,\"name\":{\"173\":{},\"214\":{}},\"comment\":{}}],[\"angle\",{\"_index\":17,\"name\":{\"18\":{},\"53\":{},\"67\":{}},\"comment\":{}}],[\"bbox\",{\"_index\":33,\"name\":{\"39\":{},\"63\":{},\"104\":{},\"138\":{},\"182\":{},\"223\":{},\"263\":{}},\"comment\":{}}],[\"body\",{\"_index\":20,\"name\":{\"22\":{},\"35\":{}},\"comment\":{}}],[\"bodymoved\",{\"_index\":108,\"name\":{\"323\":{}},\"comment\":{}}],[\"bodyoptions\",{\"_index\":13,\"name\":{\"14\":{}},\"comment\":{}}],[\"bodyprops\",{\"_index\":30,\"name\":{\"36\":{}},\"comment\":{}}],[\"bodytype\",{\"_index\":2,\"name\":{\"2\":{}},\"comment\":{}}],[\"box\",{\"_index\":6,\"name\":{\"6\":{},\"169\":{}},\"comment\":{}}],[\"center\",{\"_index\":60,\"name\":{\"98\":{}},\"comment\":{}}],[\"centered\",{\"_index\":64,\"name\":{\"111\":{},\"147\":{},\"189\":{},\"230\":{},\"270\":{}},\"comment\":{}}],[\"checkainb\",{\"_index\":111,\"name\":{\"326\":{}},\"comment\":{}}],[\"checkall\",{\"_index\":89,\"name\":{\"300\":{}},\"comment\":{}}],[\"checkcollision\",{\"_index\":91,\"name\":{\"302\":{}},\"comment\":{}}],[\"checkone\",{\"_index\":88,\"name\":{\"299\":{}},\"comment\":{}}],[\"children\",{\"_index\":11,\"name\":{\"11\":{}},\"comment\":{}}],[\"childrendata\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"circle\",{\"_index\":5,\"name\":{\"5\":{},\"57\":{}},\"comment\":{}}],[\"circleincircle\",{\"_index\":124,\"name\":{\"341\":{}},\"comment\":{}}],[\"circleinpolygon\",{\"_index\":125,\"name\":{\"342\":{}},\"comment\":{}}],[\"circleoutsidepolygon\",{\"_index\":126,\"name\":{\"343\":{}},\"comment\":{}}],[\"clockwise\",{\"_index\":106,\"name\":{\"321\":{}},\"comment\":{}}],[\"clonepointsarray\",{\"_index\":112,\"name\":{\"327\":{}},\"comment\":{}}],[\"constructor\",{\"_index\":45,\"name\":{\"58\":{},\"88\":{},\"133\":{},\"170\":{},\"211\":{},\"252\":{}},\"comment\":{}}],[\"convexpolygons\",{\"_index\":62,\"name\":{\"105\":{},\"140\":{},\"183\":{},\"224\":{},\"264\":{}},\"comment\":{}}],[\"createbox\",{\"_index\":98,\"name\":{\"311\":{},\"317\":{}},\"comment\":{}}],[\"createcircle\",{\"_index\":97,\"name\":{\"310\":{}},\"comment\":{}}],[\"createellipse\",{\"_index\":99,\"name\":{\"312\":{},\"316\":{}},\"comment\":{}}],[\"createline\",{\"_index\":96,\"name\":{\"309\":{}},\"comment\":{}}],[\"createpoint\",{\"_index\":95,\"name\":{\"308\":{}},\"comment\":{}}],[\"createpolygon\",{\"_index\":100,\"name\":{\"313\":{}},\"comment\":{}}],[\"dashlineto\",{\"_index\":113,\"name\":{\"328\":{}},\"comment\":{}}],[\"data\",{\"_index\":12,\"name\":{\"12\":{},\"13\":{},\"305\":{}},\"comment\":{}}],[\"decomppoint\",{\"_index\":1,\"name\":{\"1\":{}},\"comment\":{}}],[\"decomppolygon\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"deg2rad\",{\"_index\":101,\"name\":{\"314\":{},\"334\":{}},\"comment\":{}}],[\"distance\",{\"_index\":105,\"name\":{\"320\":{}},\"comment\":{}}],[\"draw\",{\"_index\":42,\"name\":{\"48\":{},\"85\":{},\"121\":{},\"158\":{},\"200\":{},\"241\":{},\"281\":{},\"306\":{}},\"comment\":{}}],[\"drawbvh\",{\"_index\":94,\"name\":{\"307\":{}},\"comment\":{}}],[\"drawpolygon\",{\"_index\":118,\"name\":{\"333\":{}},\"comment\":{}}],[\"ellipse\",{\"_index\":3,\"name\":{\"3\":{},\"87\":{}},\"comment\":{}}],[\"end\",{\"_index\":80,\"name\":{\"256\":{}},\"comment\":{}}],[\"ensureconvex\",{\"_index\":119,\"name\":{\"336\":{}},\"comment\":{}}],[\"ensurepolygonpoints\",{\"_index\":104,\"name\":{\"319\":{}},\"comment\":{}}],[\"ensurevectorpoint\",{\"_index\":103,\"name\":{\"318\":{}},\"comment\":{}}],[\"extendbody\",{\"_index\":107,\"name\":{\"322\":{}},\"comment\":{}}],[\"getaabbasbbox\",{\"_index\":43,\"name\":{\"49\":{},\"84\":{},\"120\":{},\"157\":{},\"199\":{},\"240\":{},\"280\":{}},\"comment\":{}}],[\"getaabbasbox\",{\"_index\":25,\"name\":{\"29\":{},\"30\":{}},\"comment\":{}}],[\"getbouncedirection\",{\"_index\":116,\"name\":{\"331\":{}},\"comment\":{}}],[\"getcentroid\",{\"_index\":81,\"name\":{\"257\":{}},\"comment\":{}}],[\"getcentroidwithoutrotation\",{\"_index\":66,\"name\":{\"122\":{},\"159\":{},\"201\":{},\"242\":{},\"282\":{}},\"comment\":{}}],[\"getconvex\",{\"_index\":72,\"name\":{\"128\":{},\"165\":{},\"207\":{},\"248\":{},\"288\":{}},\"comment\":{}}],[\"getoffsetwithangle\",{\"_index\":53,\"name\":{\"86\":{}},\"comment\":{}}],[\"getpotentials\",{\"_index\":90,\"name\":{\"301\":{}},\"comment\":{}}],[\"getsattest\",{\"_index\":117,\"name\":{\"332\":{}},\"comment\":{}}],[\"h\",{\"_index\":29,\"name\":{\"34\":{}},\"comment\":{}}],[\"height\",{\"_index\":78,\"name\":{\"176\":{},\"217\":{}},\"comment\":{}}],[\"insert\",{\"_index\":85,\"name\":{\"295\":{}},\"comment\":{}}],[\"intersectaabb\",{\"_index\":110,\"name\":{\"325\":{}},\"comment\":{}}],[\"intersectlinecircle\",{\"_index\":127,\"name\":{\"344\":{}},\"comment\":{}}],[\"intersectlineline\",{\"_index\":129,\"name\":{\"346\":{}},\"comment\":{}}],[\"intersectlinelinefast\",{\"_index\":128,\"name\":{\"345\":{}},\"comment\":{}}],[\"intersectlinepolygon\",{\"_index\":130,\"name\":{\"347\":{}},\"comment\":{}}],[\"iscentered\",{\"_index\":16,\"name\":{\"17\":{},\"52\":{},\"73\":{},\"94\":{},\"149\":{},\"191\":{},\"232\":{},\"272\":{}},\"comment\":{}}],[\"isconvex\",{\"_index\":32,\"name\":{\"38\":{},\"71\":{},\"90\":{},\"139\":{},\"172\":{},\"213\":{},\"254\":{}},\"comment\":{}}],[\"issimple\",{\"_index\":70,\"name\":{\"126\":{},\"131\":{},\"163\":{},\"205\":{},\"246\":{},\"286\":{}},\"comment\":{}}],[\"isstatic\",{\"_index\":14,\"name\":{\"15\":{},\"50\":{},\"68\":{},\"107\":{},\"142\":{},\"185\":{},\"226\":{},\"266\":{}},\"comment\":{}}],[\"istrigger\",{\"_index\":15,\"name\":{\"16\":{},\"51\":{},\"69\":{},\"108\":{},\"143\":{},\"186\":{},\"227\":{},\"267\":{}},\"comment\":{}}],[\"leaf\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"line\",{\"_index\":4,\"name\":{\"4\":{},\"251\":{}},\"comment\":{}}],[\"maparraytovector\",{\"_index\":115,\"name\":{\"330\":{}},\"comment\":{}}],[\"mapvectortoarray\",{\"_index\":114,\"name\":{\"329\":{}},\"comment\":{}}],[\"maxx\",{\"_index\":47,\"name\":{\"60\":{},\"101\":{},\"135\":{},\"179\":{},\"220\":{},\"260\":{}},\"comment\":{}}],[\"maxy\",{\"_index\":49,\"name\":{\"62\":{},\"103\":{},\"137\":{},\"181\":{},\"222\":{},\"262\":{}},\"comment\":{}}],[\"minx\",{\"_index\":46,\"name\":{\"59\":{},\"100\":{},\"134\":{},\"178\":{},\"219\":{},\"259\":{}},\"comment\":{}}],[\"miny\",{\"_index\":48,\"name\":{\"61\":{},\"102\":{},\"136\":{},\"180\":{},\"221\":{},\"261\":{}},\"comment\":{}}],[\"notintersectaabb\",{\"_index\":109,\"name\":{\"324\":{}},\"comment\":{}}],[\"offset\",{\"_index\":34,\"name\":{\"40\":{},\"64\":{}},\"comment\":{}}],[\"offsetcopy\",{\"_index\":50,\"name\":{\"65\":{}},\"comment\":{}}],[\"padding\",{\"_index\":18,\"name\":{\"19\":{},\"54\":{},\"66\":{},\"106\":{},\"141\":{},\"184\":{},\"225\":{},\"265\":{}},\"comment\":{}}],[\"point\",{\"_index\":7,\"name\":{\"7\":{},\"21\":{},\"210\":{}},\"comment\":{}}],[\"pointinpolygon\",{\"_index\":121,\"name\":{\"338\":{}},\"comment\":{}}],[\"pointoncircle\",{\"_index\":123,\"name\":{\"340\":{}},\"comment\":{}}],[\"pointsbackup\",{\"_index\":63,\"name\":{\"110\":{},\"146\":{},\"188\":{},\"229\":{},\"269\":{}},\"comment\":{}}],[\"polygon\",{\"_index\":8,\"name\":{\"8\":{},\"132\":{}},\"comment\":{}}],[\"polygonincircle\",{\"_index\":120,\"name\":{\"337\":{}},\"comment\":{}}],[\"polygoninpolygon\",{\"_index\":122,\"name\":{\"339\":{}},\"comment\":{}}],[\"pos\",{\"_index\":27,\"name\":{\"32\":{}},\"comment\":{}}],[\"potentialvector\",{\"_index\":21,\"name\":{\"23\":{}},\"comment\":{}}],[\"rad2deg\",{\"_index\":102,\"name\":{\"315\":{},\"335\":{}},\"comment\":{}}],[\"radiusx\",{\"_index\":58,\"name\":{\"96\":{}},\"comment\":{}}],[\"radiusy\",{\"_index\":59,\"name\":{\"97\":{}},\"comment\":{}}],[\"ray\",{\"_index\":83,\"name\":{\"293\":{}},\"comment\":{}}],[\"raycast\",{\"_index\":92,\"name\":{\"303\":{}},\"comment\":{}}],[\"raycasthit\",{\"_index\":19,\"name\":{\"20\":{}},\"comment\":{}}],[\"remove\",{\"_index\":84,\"name\":{\"294\":{}},\"comment\":{}}],[\"response\",{\"_index\":82,\"name\":{\"292\":{}},\"comment\":{}}],[\"rotate\",{\"_index\":69,\"name\":{\"125\":{},\"162\":{},\"204\":{},\"245\":{},\"285\":{}},\"comment\":{}}],[\"sattest\",{\"_index\":44,\"name\":{\"55\":{}},\"comment\":{}}],[\"scale\",{\"_index\":52,\"name\":{\"77\":{},\"117\":{},\"154\":{},\"196\":{},\"237\":{},\"277\":{}},\"comment\":{}}],[\"scalevector\",{\"_index\":65,\"name\":{\"112\":{},\"148\":{},\"190\":{},\"231\":{},\"271\":{}},\"comment\":{}}],[\"scalex\",{\"_index\":36,\"name\":{\"42\":{},\"78\":{},\"115\":{},\"152\":{},\"194\":{},\"235\":{},\"275\":{}},\"comment\":{}}],[\"scaley\",{\"_index\":37,\"name\":{\"43\":{},\"79\":{},\"116\":{},\"153\":{},\"195\":{},\"236\":{},\"276\":{}},\"comment\":{}}],[\"separate\",{\"_index\":87,\"name\":{\"298\":{}},\"comment\":{}}],[\"setangle\",{\"_index\":40,\"name\":{\"46\":{},\"82\":{}},\"comment\":{}}],[\"setoffset\",{\"_index\":41,\"name\":{\"47\":{},\"83\":{}},\"comment\":{}}],[\"setpoints\",{\"_index\":67,\"name\":{\"123\":{},\"160\":{},\"202\":{},\"243\":{},\"283\":{}},\"comment\":{}}],[\"setposition\",{\"_index\":38,\"name\":{\"44\":{},\"80\":{},\"118\":{},\"155\":{},\"197\":{},\"238\":{},\"278\":{}},\"comment\":{}}],[\"setscale\",{\"_index\":39,\"name\":{\"45\":{},\"81\":{},\"119\":{},\"156\":{},\"198\":{},\"239\":{},\"279\":{}},\"comment\":{}}],[\"start\",{\"_index\":79,\"name\":{\"255\":{}},\"comment\":{}}],[\"step\",{\"_index\":57,\"name\":{\"95\":{}},\"comment\":{}}],[\"system\",{\"_index\":35,\"name\":{\"41\":{},\"70\":{},\"109\":{},\"144\":{},\"187\":{},\"228\":{},\"268\":{},\"291\":{}},\"comment\":{}}],[\"translate\",{\"_index\":68,\"name\":{\"124\":{},\"161\":{},\"203\":{},\"244\":{},\"284\":{}},\"comment\":{}}],[\"traverse\",{\"_index\":93,\"name\":{\"304\":{}},\"comment\":{}}],[\"type\",{\"_index\":31,\"name\":{\"37\":{},\"72\":{},\"89\":{},\"145\":{},\"171\":{},\"212\":{},\"253\":{}},\"comment\":{}}],[\"unscaledradius\",{\"_index\":51,\"name\":{\"74\":{}},\"comment\":{}}],[\"update\",{\"_index\":86,\"name\":{\"297\":{}},\"comment\":{}}],[\"updatebody\",{\"_index\":74,\"name\":{\"130\":{},\"168\":{},\"209\":{},\"250\":{},\"290\":{},\"296\":{}},\"comment\":{}}],[\"updateconvexpolygonpositions\",{\"_index\":71,\"name\":{\"127\":{},\"164\":{},\"206\":{},\"247\":{},\"287\":{}},\"comment\":{}}],[\"updateconvexpolygons\",{\"_index\":73,\"name\":{\"129\":{},\"166\":{},\"208\":{},\"249\":{},\"289\":{}},\"comment\":{}}],[\"updateisconvex\",{\"_index\":61,\"name\":{\"99\":{},\"167\":{},\"177\":{},\"218\":{},\"258\":{}},\"comment\":{}}],[\"vector\",{\"_index\":24,\"name\":{\"26\":{}},\"comment\":{}}],[\"w\",{\"_index\":28,\"name\":{\"33\":{}},\"comment\":{}}],[\"width\",{\"_index\":77,\"name\":{\"175\":{},\"216\":{}},\"comment\":{}}],[\"x\",{\"_index\":22,\"name\":{\"24\":{},\"27\":{},\"75\":{},\"113\":{},\"150\":{},\"192\":{},\"233\":{},\"273\":{}},\"comment\":{}}],[\"y\",{\"_index\":23,\"name\":{\"25\":{},\"28\":{},\"76\":{},\"114\":{},\"151\":{},\"193\":{},\"234\":{},\"274\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
