+++ docs/functions/dashLineTo.html
@@ -40,7 +40,7 @@
 <h5>gap: <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> = 4</span></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">void</span></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L197">src/utils.ts:197</a></li></ul></aside></li></ul></section></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L195">src/utils.ts:195</a></li></ul></aside></li></ul></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
