+++ docs/functions/ensureConvex.html
@@ -25,7 +25,7 @@
 <h4>Type Parameters</h4>
 <ul class="tsd-type-parameters">
 <li>
-<h4>T<span class="tsd-signature-symbol"> extends </span><a href="../types/Body.html" class="tsd-signature-type" data-tsd-kind="Type alias">Body</a> = <a href="../classes/Polygon.html" class="tsd-signature-type" data-tsd-kind="Class">Polygon</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Circle.html" class="tsd-signature-type" data-tsd-kind="Class">Circle</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Point.html" class="tsd-signature-type" data-tsd-kind="Class">Point</a></h4></li></ul></section>
+<h4>T<span class="tsd-signature-symbol"> extends </span><a href="../types/Body.html" class="tsd-signature-type" data-tsd-kind="Type alias">Body</a> = <a href="../classes/Circle.html" class="tsd-signature-type" data-tsd-kind="Class">Circle</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Point.html" class="tsd-signature-type" data-tsd-kind="Class">Point</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Polygon.html" class="tsd-signature-type" data-tsd-kind="Class">Polygon</a></h4></li></ul></section>
 <div class="tsd-parameters">
 <h4 class="tsd-parameters-title">Parameters</h4>
 <ul class="tsd-parameter-list">
@@ -33,7 +33,7 @@
 <h5>body: <span class="tsd-signature-type" data-tsd-kind="Type parameter">T</span></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-symbol">(</span><span class="tsd-signature-type" data-tsd-kind="Type parameter">T</span><span class="tsd-signature-symbol"> | </span><a href="../classes/SATPolygon.html" class="tsd-signature-type" data-tsd-kind="Class">SATPolygon</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">[]</span></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L248">src/utils.ts:248</a></li></ul></aside></li></ul></section></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/intersect.ts#L13">src/intersect.ts:13</a></li></ul></aside></li></ul></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
