+++ docs/functions/ensureVectorPoint.html
@@ -28,7 +28,7 @@
 <h5>point: <a href="../interfaces/PotentialVector.html" class="tsd-signature-type" data-tsd-kind="Interface">PotentialVector</a><span class="tsd-signature-symbol"> = {}</span></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <a href="../classes/SATVector.html" class="tsd-signature-type" data-tsd-kind="Class">SATVector</a></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L82">src/utils.ts:82</a></li></ul></aside></li></ul></section></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L80">src/utils.ts:80</a></li></ul></aside></li></ul></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
