+++ docs/functions/getBounceDirection.html
@@ -30,7 +30,7 @@
 <h5>collider: <a href="../interfaces/Vector.html" class="tsd-signature-type" data-tsd-kind="Interface">Vector</a></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <a href="../classes/SATVector.html" class="tsd-signature-type" data-tsd-kind="Class">SATVector</a></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L261">src/utils.ts:261</a></li></ul></aside></li></ul></section></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L246">src/utils.ts:246</a></li></ul></aside></li></ul></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
