+++ docs/functions/getSATTest.html
@@ -30,7 +30,7 @@
 <h5>wall: <a href="../types/Body.html" class="tsd-signature-type" data-tsd-kind="Type alias">Body</a></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-symbol">(</span><span class="tsd-signature-symbol">(</span>a<span class="tsd-signature-symbol">: </span><a href="../classes/SATCircle.html" class="tsd-signature-type" data-tsd-kind="Class">SATCircle</a>, b<span class="tsd-signature-symbol">: </span><a href="../classes/SATCircle.html" class="tsd-signature-type" data-tsd-kind="Class">SATCircle</a>, response<span class="tsd-signature-symbol">?: </span><a href="../classes/Response.html" class="tsd-signature-type" data-tsd-kind="Class">Response</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> =&gt; </span><span class="tsd-signature-type">boolean</span><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-symbol">(</span><span class="tsd-signature-symbol">(</span>circle<span class="tsd-signature-symbol">: </span><a href="../classes/SATCircle.html" class="tsd-signature-type" data-tsd-kind="Class">SATCircle</a>, polygon<span class="tsd-signature-symbol">: </span><a href="../classes/SATPolygon.html" class="tsd-signature-type" data-tsd-kind="Class">SATPolygon</a>, response<span class="tsd-signature-symbol">?: </span><a href="../classes/Response.html" class="tsd-signature-type" data-tsd-kind="Class">Response</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> =&gt; </span><span class="tsd-signature-type">boolean</span><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-symbol">(</span><span class="tsd-signature-symbol">(</span>polygon<span class="tsd-signature-symbol">: </span><a href="../classes/SATPolygon.html" class="tsd-signature-type" data-tsd-kind="Class">SATPolygon</a>, circle<span class="tsd-signature-symbol">: </span><a href="../classes/SATCircle.html" class="tsd-signature-type" data-tsd-kind="Class">SATCircle</a>, response<span class="tsd-signature-symbol">?: </span><a href="../classes/Response.html" class="tsd-signature-type" data-tsd-kind="Class">Response</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> =&gt; </span><span class="tsd-signature-type">boolean</span><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-symbol">(</span><span class="tsd-signature-symbol">(</span>a<span class="tsd-signature-symbol">: </span><a href="../classes/SATPolygon.html" class="tsd-signature-type" data-tsd-kind="Class">SATPolygon</a>, b<span class="tsd-signature-symbol">: </span><a href="../classes/SATPolygon.html" class="tsd-signature-type" data-tsd-kind="Class">SATPolygon</a>, response<span class="tsd-signature-symbol">?: </span><a href="../classes/Response.html" class="tsd-signature-type" data-tsd-kind="Class">Response</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> =&gt; </span><span class="tsd-signature-type">boolean</span><span class="tsd-signature-symbol">)</span></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L272">src/utils.ts:272</a></li></ul></aside></li></ul></section></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L257">src/utils.ts:257</a></li></ul></aside></li></ul></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
