+++ docs/functions/intersectAABB.html
@@ -30,7 +30,7 @@
 <h5>b: <a href="../interfaces/BBox.html" class="tsd-signature-type" data-tsd-kind="Interface">BBox</a></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">boolean</span></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L161">src/utils.ts:161</a></li></ul></aside></li></ul></section></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L159">src/utils.ts:159</a></li></ul></aside></li></ul></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
