+++ docs/functions/pointOnCircle.html
@@ -30,7 +30,7 @@
 <h5>circle: <span class="tsd-signature-type">Pick</span><span class="tsd-signature-symbol">&lt;</span><a href="../classes/Circle.html" class="tsd-signature-type" data-tsd-kind="Class">Circle</a><span class="tsd-signature-symbol">, </span><span class="tsd-signature-type">&quot;pos&quot;</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">&quot;r&quot;</span><span class="tsd-signature-symbol">&gt;</span></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">boolean</span></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/intersect.ts#L43">src/intersect.ts:43</a></li></ul></aside></li></ul></section></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/intersect.ts#L56">src/intersect.ts:56</a></li></ul></aside></li></ul></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
