+++ docs/functions/polygonInPolygon.html
@@ -28,7 +28,7 @@
 <h5>polygonB: <a href="../classes/Polygon.html" class="tsd-signature-type" data-tsd-kind="Class">Polygon</a></h5></li></ul></div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">boolean</span></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/intersect.ts#L28">src/intersect.ts:28</a></li></ul></aside></li></ul></section></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/intersect.ts#L41">src/intersect.ts:41</a></li></ul></aside></li></ul></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
