+++ docs/variables/RAD2DEG.html
@@ -17,7 +17,7 @@
 <h1>Variable RAD2DEG<code class="tsd-tag ts-flagConst">Const</code> </h1></div>
 <div class="tsd-signature">RAD2DEG<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> = ...</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L31">src/utils.ts:31</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/Prozi/detect-collisions/blob/master/src/utils.ts#L29">src/utils.ts:29</a></li></ul></aside></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
